Move serving to core
