+++ examples/tutorials/flow-deploy/distribute-flow-as-executable-app/main.py
@@ -2,14 +2,15 @@ import base64
 import json
 import os
 import re
-import streamlit as st
 from pathlib import Path
-from streamlit_quill import st_quill  # noqa: F401
+
+import streamlit as st
 from bs4 import BeautifulSoup, NavigableString, Tag
+from streamlit_quill import st_quill  # noqa: F401
 
 from promptflow._sdk._utils import print_yellow_warning
-from promptflow._sdk._serving.flow_invoker import FlowInvoker
-from promptflow._utils.multimedia_utils import is_multimedia_dict, MIME_PATTERN
+from promptflow._utils.multimedia_utils import MIME_PATTERN, is_multimedia_dict
+from promptflow.core._serving.flow_invoker import FlowInvoker
 
 invoker = None
 
@@ -20,7 +21,7 @@ def start():
 
     def show_image(image, key=None):
         if not image.startswith("data:image"):
-            st.image(key + ',' + image)
+            st.image(key + "," + image)
         else:
             st.image(image)
 
@@ -50,12 +51,12 @@ def start():
             elif isinstance(rich_text_value, dict):
                 result |= is_dict_contains_rich_text(rich_text_value)
             elif re.match(MIME_PATTERN, rich_text_key) or (
-                    isinstance(rich_text_value, str) and rich_text_value.startswith("data:image")):
+                isinstance(rich_text_value, str) and rich_text_value.startswith("data:image")
+            ):
                 result = True
         return result
 
     def render_message(role, message_items):
-
         def item_render_message(value, key=None):
             if key and re.match(MIME_PATTERN, key):
                 show_image(value, key)
@@ -154,8 +155,8 @@ def start():
             if text:
                 return [text]
         elif isinstance(node, Tag):
-            if node.name == 'img':
-                prefix, base64_str = node['src'].split(',', 1)
+            if node.name == "img":
+                prefix, base64_str = node["src"].split(",", 1)
                 return [{prefix: base64_str}]
             else:
                 result = []
@@ -165,16 +166,16 @@ def start():
         return []
 
     def parse_html_content(html_content):
-        soup = BeautifulSoup(html_content, 'html.parser')
+        soup = BeautifulSoup(html_content, "html.parser")
         result = []
-        for p in soup.find_all('p'):
+        for p in soup.find_all("p"):
             result.extend(extract_content(p))
         return result
 
     def parse_image_content(image_content, image_type):
         if image_content is not None:
             file_contents = image_content.read()
-            image_content = base64.b64encode(file_contents).decode('utf-8')
+            image_content = base64.b64encode(file_contents).decode("utf-8")
             prefix = f"data:{image_type};base64"
             return {prefix: image_content}
 
@@ -184,7 +185,7 @@ def start():
     with container:
         show_conversation()
 
-    with st.form(key='input_form', clear_on_submit=True):
+    with st.form(key="input_form", clear_on_submit=True):
         settings_path = os.path.join(os.path.dirname(__file__), "settings.json")
         if os.path.exists(settings_path):
             with open(settings_path, "r") as file:
@@ -195,15 +196,17 @@ def start():
                     label=environment_variable,
                     type="password",
                     placeholder=f"Please input {environment_variable} here. If you input before, you can leave it "
-                                f"blank.")
+                    f"blank.",
+                )
                 if secret_input != "":
                     os.environ[environment_variable] = secret_input
 
-        url = st.text_input(label='url',
-                            placeholder='https://play.google.com/store/apps/details?id=com.twitter.android')
+        url = st.text_input(
+            label="url", placeholder="https://play.google.com/store/apps/details?id=com.twitter.android"
+        )
         cols = st.columns(7)
-        submit_bt = cols[0].form_submit_button(label='Submit')
-        clear_bt = cols[1].form_submit_button(label='Clear')
+        submit_bt = cols[0].form_submit_button(label="Submit")
+        clear_bt = cols[1].form_submit_button(label="Clear")
 
     if submit_bt:
         submit(url=url)
