+++ src/promptflow/promptflow/_internal/__init__.py
@@ -42,16 +42,6 @@ from promptflow._core.tools_manager import (
     retrieve_tool_func_result,
 )
 from promptflow._sdk._constants import LOCAL_MGMT_DB_PATH
-from promptflow._sdk._serving.response_creator import ResponseCreator
-from promptflow._sdk._serving.swagger import generate_swagger
-from promptflow._sdk._serving.utils import (
-    get_output_fields_to_remove,
-    get_sample_json,
-    handle_error_to_response,
-    load_request_data,
-    streaming_response_required,
-    validate_request_data,
-)
 from promptflow._sdk._utils import (
     get_used_connection_names_from_environment_variables,
     setup_user_agent_to_operation_context,
@@ -107,6 +97,16 @@ from promptflow._utils.utils import (
 )
 from promptflow._version import VERSION
 from promptflow.batch._csharp_base_executor_proxy import CSharpBaseExecutorProxy
+from promptflow.core._serving.response_creator import ResponseCreator
+from promptflow.core._serving.swagger import generate_swagger
+from promptflow.core._serving.utils import (
+    get_output_fields_to_remove,
+    get_sample_json,
+    handle_error_to_response,
+    load_request_data,
+    streaming_response_required,
+    validate_request_data,
+)
 from promptflow.executor._errors import InputNotFound
 from promptflow.executor._tool_invoker import DefaultToolInvoker
 from promptflow.storage._run_storage import DefaultRunStorage
