+++ src/promptflow/promptflow/core/_serving/static/index.js
@@ -29,7 +29,7 @@ INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES WHATSOEVER RESULTING FROM
 LOSS OF USE, DATA OR PROFITS, WHETHER IN AN ACTION OF CONTRACT, NEGLIGENCE OR
 OTHER TORTIOUS ACTION, ARISING OUT OF OR IN CONNECTION WITH THE USE OR
 PERFORMANCE OF THIS SOFTWARE.
-***************************************************************************** */var mE=function(e,t){return mE=Object.setPrototypeOf||{__proto__:[]}instanceof Array&&function(n,r){n.__proto__=r}||function(n,r){for(var o in r)Object.prototype.hasOwnProperty.call(r,o)&&(n[o]=r[o])},mE(e,t)};function yi(e,t){if(typeof t!="function"&&t!==null)throw new TypeError("Class extends value "+String(t)+" is not a constructor or null");mE(e,t);function n(){this.constructor=e}e.prototype=t===null?Object.create(t):(n.prototype=t.prototype,new n)}var oe=function(){return oe=Object.assign||function(t){for(var n,r=1,o=arguments.length;r<o;r++){n=arguments[r];for(var i in n)Object.prototype.hasOwnProperty.call(n,i)&&(t[i]=n[i])}return t},oe.apply(this,arguments)};function pl(e,t){var n={};for(var r in e)Object.prototype.hasOwnProperty.call(e,r)&&t.indexOf(r)<0&&(n[r]=e[r]);if(e!=null&&typeof Object.getOwnPropertySymbols=="function")for(var o=0,r=Object.getOwnPropertySymbols(e);o<r.length;o++)t.indexOf(r[o])<0&&Object.prototype.propertyIsEnumerable.call(e,r[o])&&(n[r[o]]=e[r[o]]);return n}function vG(e,t,n,r){var o=arguments.length,i=o<3?t:r===null?r=Object.getOwnPropertyDescriptor(t,n):r,a;if(typeof Reflect=="object"&&typeof Reflect.decorate=="function")i=Reflect.decorate(e,t,n,r);else for(var s=e.length-1;s>=0;s--)(a=e[s])&&(i=(o<3?a(i):o>3?a(t,n,i):a(t,n))||i);return o>3&&i&&Object.defineProperty(t,n,i),i}function Yi(e,t,n){if(n||arguments.length===2)for(var r=0,o=t.length,i;r<o;r++)(i||!(r in t))&&(i||(i=Array.prototype.slice.call(t,0,r)),i[r]=t[r]);return e.concat(i||t)}var K8={},bg=void 0;try{bg=window}catch{}function yb(e,t){if(typeof bg<"u"){var n=bg.__packages__=bg.__packages__||{};if(!n[e]||!K8[e]){K8[e]=t;var r=n[e]=n[e]||[];r.push(t)}}}yb("@fluentui/set-version","6.0.0");var j1={none:0,insertNode:1,appendChild:2},V8="__stylesheet__",bG=typeof navigator<"u"&&/rv:11.0/.test(navigator.userAgent),Lf={};try{Lf=window||{}}catch{}var vf,Qi=function(){function e(t,n){var r,o,i,a,s,l;this._rules=[],this._preservedRules=[],this._counter=0,this._keyToClassName={},this._onInsertRuleCallbacks=[],this._onResetCallbacks=[],this._classNameToArgs={},this._config=oe({injectionMode:typeof document>"u"?j1.none:j1.insertNode,defaultPrefix:"css",namespace:void 0,cspSettings:void 0},t),this._classNameToArgs=(r=n==null?void 0:n.classNameToArgs)!==null&&r!==void 0?r:this._classNameToArgs,this._counter=(o=n==null?void 0:n.counter)!==null&&o!==void 0?o:this._counter,this._keyToClassName=(a=(i=this._config.classNameCache)!==null&&i!==void 0?i:n==null?void 0:n.keyToClassName)!==null&&a!==void 0?a:this._keyToClassName,this._preservedRules=(s=n==null?void 0:n.preservedRules)!==null&&s!==void 0?s:this._preservedRules,this._rules=(l=n==null?void 0:n.rules)!==null&&l!==void 0?l:this._rules}return e.getInstance=function(){if(vf=Lf[V8],!vf||vf._lastStyleElement&&vf._lastStyleElement.ownerDocument!==document){var t=(Lf==null?void 0:Lf.FabricConfig)||{},n=new e(t.mergeStyles,t.serializedStylesheet);vf=n,Lf[V8]=n}return vf},e.prototype.serialize=function(){return JSON.stringify({classNameToArgs:this._classNameToArgs,counter:this._counter,keyToClassName:this._keyToClassName,preservedRules:this._preservedRules,rules:this._rules})},e.prototype.setConfig=function(t){this._config=oe(oe({},this._config),t)},e.prototype.onReset=function(t){var n=this;return this._onResetCallbacks.push(t),function(){n._onResetCallbacks=n._onResetCallbacks.filter(function(r){return r!==t})}},e.prototype.onInsertRule=function(t){var n=this;return this._onInsertRuleCallbacks.push(t),function(){n._onInsertRuleCallbacks=n._onInsertRuleCallbacks.filter(function(r){return r!==t})}},e.prototype.getClassName=function(t){var n=this._config.namespace,r=t||this._config.defaultPrefix;return(n?n+"-":"")+r+"-"+this._counter++},e.prototype.cacheClassName=function(t,n,r,o){this._keyToClassName[n]=t,this._classNameToArgs[t]={args:r,rules:o}},e.prototype.classNameFromKey=function(t){return this._keyToClassName[t]},e.prototype.getClassNameCache=function(){return this._keyToClassName},e.prototype.argsFromClassName=function(t){var n=this._classNameToArgs[t];return n&&n.args},e.prototype.insertedRulesFromClassName=function(t){var n=this._classNameToArgs[t];return n&&n.rules},e.prototype.insertRule=function(t,n){var r=this._config.injectionMode,o=r!==j1.none?this._getStyleElement():void 0;if(n&&this._preservedRules.push(t),o)switch(r){case j1.insertNode:var i=o.sheet;try{i.insertRule(t,i.cssRules.length)}catch{}break;case j1.appendChild:o.appendChild(document.createTextNode(t));break}else this._rules.push(t);this._config.onInsertRule&&this._config.onInsertRule(t),this._onInsertRuleCallbacks.forEach(function(a){return a()})},e.prototype.getRules=function(t){return(t?this._preservedRules.join(""):"")+this._rules.join("")},e.prototype.reset=function(){this._rules=[],this._counter=0,this._classNameToArgs={},this._keyToClassName={},this._onResetCallbacks.forEach(function(t){return t()})},e.prototype.resetKeys=function(){this._keyToClassName={}},e.prototype._getStyleElement=function(){var t=this;return!this._styleElement&&typeof document<"u"&&(this._styleElement=this._createStyleElement(),bG||window.requestAnimationFrame(function(){t._styleElement=void 0})),this._styleElement},e.prototype._createStyleElement=function(){var t=document.head,n=document.createElement("style"),r=null;n.setAttribute("data-merge-styles","true");var o=this._config.cspSettings;if(o&&o.nonce&&n.setAttribute("nonce",o.nonce),this._lastStyleElement)r=this._lastStyleElement.nextElementSibling;else{var i=this._findPlaceholderStyleTag();i?r=i.nextElementSibling:r=t.childNodes[0]}return t.insertBefore(n,t.contains(r)?r:null),this._lastStyleElement=n,n},e.prototype._findPlaceholderStyleTag=function(){var t=document.head;return t?t.querySelector("style[data-merge-styles]"):null},e}();function mF(){for(var e=[],t=0;t<arguments.length;t++)e[t]=arguments[t];var n=[],r=[],o=Qi.getInstance();function i(a){for(var s=0,l=a;s<l.length;s++){var u=l[s];if(u)if(typeof u=="string")if(u.indexOf(" ")>=0)i(u.split(" "));else{var d=o.argsFromClassName(u);d?i(d):n.indexOf(u)===-1&&n.push(u)}else Array.isArray(u)?i(u):typeof u=="object"&&r.push(u)}}return i(e),{classes:n,objects:r}}function gF(e){cd!==e&&(cd=e)}function vF(){return cd===void 0&&(cd=typeof document<"u"&&!!document.documentElement&&document.documentElement.getAttribute("dir")==="rtl"),cd}var cd;cd=vF();function Eb(){return{rtl:vF()}}var Y8={};function yG(e,t){var n=e[t];n.charAt(0)!=="-"&&(e[t]=Y8[n]=Y8[n]||n.replace(/([A-Z])/g,"-$1").toLowerCase())}var vm;function EG(){var e;if(!vm){var t=typeof document<"u"?document:void 0,n=typeof navigator<"u"?navigator:void 0,r=(e=n==null?void 0:n.userAgent)===null||e===void 0?void 0:e.toLowerCase();t?vm={isWebkit:!!(t&&"WebkitAppearance"in t.documentElement.style),isMoz:!!(r&&r.indexOf("firefox")>-1),isOpera:!!(r&&r.indexOf("opera")>-1),isMs:!!(n&&(/rv:11.0/i.test(n.userAgent)||/Edge\/\d./i.test(navigator.userAgent)))}:vm={isWebkit:!0,isMoz:!0,isOpera:!0,isMs:!0}}return vm}var Q8={"user-select":1};function _G(e,t){var n=EG(),r=e[t];if(Q8[r]){var o=e[t+1];Q8[r]&&(n.isWebkit&&e.push("-webkit-"+r,o),n.isMoz&&e.push("-moz-"+r,o),n.isMs&&e.push("-ms-"+r,o),n.isOpera&&e.push("-o-"+r,o))}}var TG=["column-count","font-weight","flex","flex-grow","flex-shrink","fill-opacity","opacity","order","z-index","zoom"];function wG(e,t){var n=e[t],r=e[t+1];if(typeof r=="number"){var o=TG.indexOf(n)>-1,i=n.indexOf("--")>-1,a=o||i?"":"px";e[t+1]=""+r+a}}var bm,Wl="left",Gl="right",kG="@noflip",X8=(bm={},bm[Wl]=Gl,bm[Gl]=Wl,bm),Z8={"w-resize":"e-resize","sw-resize":"se-resize","nw-resize":"ne-resize"};function SG(e,t,n){if(e.rtl){var r=t[n];if(!r)return;var o=t[n+1];if(typeof o=="string"&&o.indexOf(kG)>=0)t[n+1]=o.replace(/\s*(?:\/\*\s*)?\@noflip\b(?:\s*\*\/)?\s*?/g,"");else if(r.indexOf(Wl)>=0)t[n]=r.replace(Wl,Gl);else if(r.indexOf(Gl)>=0)t[n]=r.replace(Gl,Wl);else if(String(o).indexOf(Wl)>=0)t[n+1]=o.replace(Wl,Gl);else if(String(o).indexOf(Gl)>=0)t[n+1]=o.replace(Gl,Wl);else if(X8[r])t[n]=X8[r];else if(Z8[o])t[n+1]=Z8[o];else switch(r){case"margin":case"padding":t[n+1]=CG(o);break;case"box-shadow":t[n+1]=xG(o,0);break}}}function xG(e,t){var n=e.split(" "),r=parseInt(n[t],10);return n[0]=n[0].replace(String(r),String(r*-1)),n.join(" ")}function CG(e){if(typeof e=="string"){var t=e.split(" ");if(t.length===4)return t[0]+" "+t[3]+" "+t[2]+" "+t[1]}return e}function AG(e){for(var t=[],n=0,r=0,o=0;o<e.length;o++)switch(e[o]){case"(":r++;break;case")":r&&r--;break;case"	":case" ":r||(o>n&&t.push(e.substring(n,o)),n=o+1);break}return n<e.length&&t.push(e.substring(n)),t}var NG="displayName";function FG(e){var t=e&&e["&"];return t?t.displayName:void 0}var bF=/\:global\((.+?)\)/g;function IG(e){if(!bF.test(e))return e;for(var t=[],n=/\:global\((.+?)\)/g,r=null;r=n.exec(e);)r[1].indexOf(",")>-1&&t.push([r.index,r.index+r[0].length,r[1].split(",").map(function(o){return":global("+o.trim()+")"}).join(", ")]);return t.reverse().reduce(function(o,i){var a=i[0],s=i[1],l=i[2],u=o.slice(0,a),d=o.slice(s);return u+l+d},e)}function J8(e,t){return e.indexOf(":global(")>=0?e.replace(bF,"$1"):e.indexOf(":")===0?t+e:e.indexOf("&")<0?t+" "+e:e}function ex(e,t,n,r){t===void 0&&(t={__order:[]}),n.indexOf("@")===0?(n=n+"{"+e,fd([r],t,n)):n.indexOf(",")>-1?IG(n).split(",").map(function(o){return o.trim()}).forEach(function(o){return fd([r],t,J8(o,e))}):fd([r],t,J8(n,e))}function fd(e,t,n){t===void 0&&(t={__order:[]}),n===void 0&&(n="&");var r=Qi.getInstance(),o=t[n];o||(o={},t[n]=o,t.__order.push(n));for(var i=0,a=e;i<a.length;i++){var s=a[i];if(typeof s=="string"){var l=r.argsFromClassName(s);l&&fd(l,t,n)}else if(Array.isArray(s))fd(s,t,n);else for(var u in s)if(s.hasOwnProperty(u)){var d=s[u];if(u==="selectors"){var h=s.selectors;for(var p in h)h.hasOwnProperty(p)&&ex(n,t,p,h[p])}else typeof d=="object"?d!==null&&ex(n,t,u,d):d!==void 0&&(u==="margin"||u==="padding"?BG(o,u,d):o[u]=d)}}return t}function BG(e,t,n){var r=typeof n=="string"?AG(n):[n];r.length===0&&r.push(n),r[r.length-1]==="!important"&&(r=r.slice(0,-1).map(function(o){return o+" !important"})),e[t+"Top"]=r[0],e[t+"Right"]=r[1]||r[0],e[t+"Bottom"]=r[2]||r[0],e[t+"Left"]=r[3]||r[1]||r[0]}function RG(e,t){for(var n=[e.rtl?"rtl":"ltr"],r=!1,o=0,i=t.__order;o<i.length;o++){var a=i[o];n.push(a);var s=t[a];for(var l in s)s.hasOwnProperty(l)&&s[l]!==void 0&&(r=!0,n.push(l,s[l]))}return r?n.join(""):void 0}function yF(e,t){return t<=0?"":t===1?e:e+yF(e,t-1)}function _4(e,t){if(!t)return"";var n=[];for(var r in t)t.hasOwnProperty(r)&&r!==NG&&t[r]!==void 0&&n.push(r,t[r]);for(var o=0;o<n.length;o+=2)yG(n,o),wG(n,o),SG(e,n,o),_G(n,o);for(var o=1;o<n.length;o+=4)n.splice(o,1,":",n[o],";");return n.join("")}function EF(e){for(var t=[],n=1;n<arguments.length;n++)t[n-1]=arguments[n];var r=fd(t),o=RG(e,r);if(o){var i=Qi.getInstance(),a={className:i.classNameFromKey(o),key:o,args:t};if(!a.className){a.className=i.getClassName(FG(r));for(var s=[],l=0,u=r.__order;l<u.length;l++){var d=u[l];s.push(d,_4(e,r[d]))}a.rulesToInsert=s}return a}}function _F(e,t){t===void 0&&(t=1);var n=Qi.getInstance(),r=e.className,o=e.key,i=e.args,a=e.rulesToInsert;if(a){for(var s=0;s<a.length;s+=2){var l=a[s+1];if(l){var u=a[s];u=u.replace(/&/g,yF("."+e.className,t));var d=u+"{"+l+"}"+(u.indexOf("@")===0?"}":"");n.insertRule(d)}}n.cacheClassName(r,o,i,a)}}function OG(e){for(var t=[],n=1;n<arguments.length;n++)t[n-1]=arguments[n];var r=EF.apply(void 0,Yi([e],t));return r?(_F(r,e.specificityMultiplier),r.className):""}function Fo(){for(var e=[],t=0;t<arguments.length;t++)e[t]=arguments[t];return TF(e,Eb())}function TF(e,t){var n=e instanceof Array?e:[e],r=mF(n),o=r.classes,i=r.objects;return i.length&&o.push(OG(t||{},i)),o.join(" ")}function jc(){for(var e=[],t=0;t<arguments.length;t++)e[t]=arguments[t];if(e&&e.length===1&&e[0]&&!e[0].subComponentStyles)return e[0];for(var n={},r={},o=0,i=e;o<i.length;o++){var a=i[o];if(a){for(var s in a)if(a.hasOwnProperty(s)){if(s==="subComponentStyles"&&a.subComponentStyles!==void 0){var l=a.subComponentStyles;for(var u in l)l.hasOwnProperty(u)&&(r.hasOwnProperty(u)?r[u].push(l[u]):r[u]=[l[u]]);continue}var d=n[s],h=a[s];d===void 0?n[s]=h:n[s]=Yi(Yi([],Array.isArray(d)?d:[d]),Array.isArray(h)?h:[h])}}}if(Object.keys(r).length>0){n.subComponentStyles={};var p=n.subComponentStyles,m=function(v){if(r.hasOwnProperty(v)){var _=r[v];p[v]=function(b){return jc.apply(void 0,_.map(function(E){return typeof E=="function"?E(b):E}))}}};for(var u in r)m(u)}return n}function q0(){for(var e=[],t=0;t<arguments.length;t++)e[t]=arguments[t];return wF(e,Eb())}function wF(e,t){var n={subComponentStyles:{}},r=e[0];if(!r&&e.length<=1)return{subComponentStyles:{}};var o=jc.apply(void 0,e),i=[];for(var a in o)if(o.hasOwnProperty(a)){if(a==="subComponentStyles"){n.subComponentStyles=o.subComponentStyles||{};continue}var s=o[a],l=mF(s),u=l.classes,d=l.objects;if(d!=null&&d.length){var h=EF(t||{},{displayName:a},d);h&&(i.push(h),n[a]=u.concat([h.className]).join(" "))}else n[a]=u.join(" ")}for(var p=0,m=i;p<m.length;p++){var h=m[p];h&&_F(h,t==null?void 0:t.specificityMultiplier)}return n}function kF(e){for(var t=[],n=1;n<arguments.length;n++)t[n-1]=arguments[n];for(var r=[],o=0,i=t;o<i.length;o++){var a=i[o];a&&r.push(typeof a=="function"?a(e):a)}return r.length===1?r[0]:r.length?jc.apply(void 0,r):{}}function SF(e){var t=Qi.getInstance(),n=_4(Eb(),e),r=t.classNameFromKey(n);if(!r){var o=t.getClassName();t.insertRule("@font-face{"+n+"}",!0),t.cacheClassName(o,n,[],["font-face",n])}}function Ji(e){var t=Qi.getInstance(),n=[];for(var r in e)e.hasOwnProperty(r)&&n.push(r,"{",_4(Eb(),e[r]),"}");var o=n.join(""),i=t.classNameFromKey(o);if(i)return i;var a=t.getClassName();return t.insertRule("@keyframes "+a+"{"+o+"}",!0),t.cacheClassName(a,o,[],["keyframes",o]),a}function DG(e){var t={},n=function(o){if(e.hasOwnProperty(o)){var i;Object.defineProperty(t,o,{get:function(){return i===void 0&&(i=Fo(e[o]).toString()),i},enumerable:!0,configurable:!0})}};for(var r in e)n(r);return t}var gE=void 0;try{gE=window}catch{}function ur(e){if(!(typeof gE>"u")){var t=e;return t&&t.ownerDocument&&t.ownerDocument.defaultView?t.ownerDocument.defaultView:gE}}var _b=function(){function e(t,n){this._timeoutIds=null,this._immediateIds=null,this._intervalIds=null,this._animationFrameIds=null,this._isDisposed=!1,this._parent=t||null,this._onErrorHandler=n,this._noop=function(){}}return e.prototype.dispose=function(){var t;if(this._isDisposed=!0,this._parent=null,this._timeoutIds){for(t in this._timeoutIds)this._timeoutIds.hasOwnProperty(t)&&this.clearTimeout(parseInt(t,10));this._timeoutIds=null}if(this._immediateIds){for(t in this._immediateIds)this._immediateIds.hasOwnProperty(t)&&this.clearImmediate(parseInt(t,10));this._immediateIds=null}if(this._intervalIds){for(t in this._intervalIds)this._intervalIds.hasOwnProperty(t)&&this.clearInterval(parseInt(t,10));this._intervalIds=null}if(this._animationFrameIds){for(t in this._animationFrameIds)this._animationFrameIds.hasOwnProperty(t)&&this.cancelAnimationFrame(parseInt(t,10));this._animationFrameIds=null}},e.prototype.setTimeout=function(t,n){var r=this,o=0;return this._isDisposed||(this._timeoutIds||(this._timeoutIds={}),o=setTimeout(function(){try{r._timeoutIds&&delete r._timeoutIds[o],t.apply(r._parent)}catch(i){r._logError(i)}},n),this._timeoutIds[o]=!0),o},e.prototype.clearTimeout=function(t){this._timeoutIds&&this._timeoutIds[t]&&(clearTimeout(t),delete this._timeoutIds[t])},e.prototype.setImmediate=function(t,n){var r=this,o=0,i=ur(n);if(!this._isDisposed){this._immediateIds||(this._immediateIds={});var a=function(){try{r._immediateIds&&delete r._immediateIds[o],t.apply(r._parent)}catch(s){r._logError(s)}};o=i.setTimeout(a,0),this._immediateIds[o]=!0}return o},e.prototype.clearImmediate=function(t,n){var r=ur(n);this._immediateIds&&this._immediateIds[t]&&(r.clearTimeout(t),delete this._immediateIds[t])},e.prototype.setInterval=function(t,n){var r=this,o=0;return this._isDisposed||(this._intervalIds||(this._intervalIds={}),o=setInterval(function(){try{t.apply(r._parent)}catch(i){r._logError(i)}},n),this._intervalIds[o]=!0),o},e.prototype.clearInterval=function(t){this._intervalIds&&this._intervalIds[t]&&(clearInterval(t),delete this._intervalIds[t])},e.prototype.throttle=function(t,n,r){var o=this;if(this._isDisposed)return this._noop;var i=n||0,a=!0,s=!0,l=0,u,d,h=null;r&&typeof r.leading=="boolean"&&(a=r.leading),r&&typeof r.trailing=="boolean"&&(s=r.trailing);var p=function(v){var _=Date.now(),b=_-l,E=a?i-b:i;return b>=i&&(!v||a)?(l=_,h&&(o.clearTimeout(h),h=null),u=t.apply(o._parent,d)):h===null&&s&&(h=o.setTimeout(p,E)),u},m=function(){for(var v=[],_=0;_<arguments.length;_++)v[_]=arguments[_];return d=v,p(!0)};return m},e.prototype.debounce=function(t,n,r){var o=this;if(this._isDisposed){var i=function(){};return i.cancel=function(){},i.flush=function(){return null},i.pending=function(){return!1},i}var a=n||0,s=!1,l=!0,u=null,d=0,h=Date.now(),p,m,v=null;r&&typeof r.leading=="boolean"&&(s=r.leading),r&&typeof r.trailing=="boolean"&&(l=r.trailing),r&&typeof r.maxWait=="number"&&!isNaN(r.maxWait)&&(u=r.maxWait);var _=function(C){v&&(o.clearTimeout(v),v=null),h=C},b=function(C){_(C),p=t.apply(o._parent,m)},E=function(C){var A=Date.now(),P=!1;C&&(s&&A-d>=a&&(P=!0),d=A);var I=A-d,j=a-I,H=A-h,K=!1;return u!==null&&(H>=u&&v?K=!0:j=Math.min(j,u-H)),I>=a||K||P?b(A):(v===null||!C)&&l&&(v=o.setTimeout(E,j)),p},w=function(){return!!v},k=function(){w()&&_(Date.now())},y=function(){return w()&&b(Date.now()),p},F=function(){for(var C=[],A=0;A<arguments.length;A++)C[A]=arguments[A];return m=C,E(!0)};return F.cancel=k,F.flush=y,F.pending=w,F},e.prototype.requestAnimationFrame=function(t,n){var r=this,o=0,i=ur(n);if(!this._isDisposed){this._animationFrameIds||(this._animationFrameIds={});var a=function(){try{r._animationFrameIds&&delete r._animationFrameIds[o],t.apply(r._parent)}catch(s){r._logError(s)}};o=i.requestAnimationFrame?i.requestAnimationFrame(a):i.setTimeout(a,0),this._animationFrameIds[o]=!0}return o},e.prototype.cancelAnimationFrame=function(t,n){var r=ur(n);this._animationFrameIds&&this._animationFrameIds[t]&&(r.cancelAnimationFrame?r.cancelAnimationFrame(t):r.clearTimeout(t),delete this._animationFrameIds[t])},e.prototype._logError=function(t){this._onErrorHandler&&this._onErrorHandler(t)},e}();function T4(e,t){for(var n in e)if(e.hasOwnProperty(n)&&(!t.hasOwnProperty(n)||t[n]!==e[n]))return!1;for(var n in t)if(t.hasOwnProperty(n)&&!e.hasOwnProperty(n))return!1;return!0}function pc(e){for(var t=[],n=1;n<arguments.length;n++)t[n-1]=arguments[n];return PG.apply(this,[null,e].concat(t))}function PG(e,t){for(var n=[],r=2;r<arguments.length;r++)n[r-2]=arguments[r];t=t||{};for(var o=0,i=n;o<i.length;o++){var a=i[o];if(a)for(var s in a)a.hasOwnProperty(s)&&(!e||e(s))&&(t[s]=a[s])}return t}var Ws=function(){function e(t){this._id=e._uniqueId++,this._parent=t,this._eventRecords=[]}return e.raise=function(t,n,r,o){var i;if(e._isElement(t)){if(typeof document<"u"&&document.createEvent){var a=document.createEvent("HTMLEvents");a.initEvent(n,o||!1,!0),pc(a,r),i=t.dispatchEvent(a)}else if(typeof document<"u"&&document.createEventObject){var s=document.createEventObject(r);t.fireEvent("on"+n,s)}}else for(;t&&i!==!1;){var l=t.__events__,u=l?l[n]:null;if(u){for(var d in u)if(u.hasOwnProperty(d))for(var h=u[d],p=0;i!==!1&&p<h.length;p++){var m=h[p];m.objectCallback&&(i=m.objectCallback.call(m.parent,r))}}t=o?t.parent:null}return i},e.isObserved=function(t,n){var r=t&&t.__events__;return!!r&&!!r[n]},e.isDeclared=function(t,n){var r=t&&t.__declaredEvents;return!!r&&!!r[n]},e.stopPropagation=function(t){t.stopPropagation?t.stopPropagation():t.cancelBubble=!0},e._isElement=function(t){return!!t&&(!!t.addEventListener||typeof HTMLElement<"u"&&t instanceof HTMLElement)},e.prototype.dispose=function(){this._isDisposed||(this._isDisposed=!0,this.off(),this._parent=null)},e.prototype.onAll=function(t,n,r){for(var o in n)n.hasOwnProperty(o)&&this.on(t,o,n[o],r)},e.prototype.on=function(t,n,r,o){var i=this;if(n.indexOf(",")>-1)for(var a=n.split(/[ ,]+/),s=0;s<a.length;s++)this.on(t,a[s],r,o);else{var l=this._parent,u={target:t,eventName:n,parent:l,callback:r,options:o},a=t.__events__=t.__events__||{};if(a[n]=a[n]||{count:0},a[n][this._id]=a[n][this._id]||[],a[n][this._id].push(u),a[n].count++,e._isElement(t)){var d=function(){for(var m=[],v=0;v<arguments.length;v++)m[v]=arguments[v];if(!i._isDisposed){var _;try{if(_=r.apply(l,m),_===!1&&m[0]){var b=m[0];b.preventDefault&&b.preventDefault(),b.stopPropagation&&b.stopPropagation(),b.cancelBubble=!0}}catch{}return _}};u.elementCallback=d,t.addEventListener?t.addEventListener(n,d,o):t.attachEvent&&t.attachEvent("on"+n,d)}else{var h=function(){for(var m=[],v=0;v<arguments.length;v++)m[v]=arguments[v];if(!i._isDisposed)return r.apply(l,m)};u.objectCallback=h}this._eventRecords.push(u)}},e.prototype.off=function(t,n,r,o){for(var i=0;i<this._eventRecords.length;i++){var a=this._eventRecords[i];if((!t||t===a.target)&&(!n||n===a.eventName)&&(!r||r===a.callback)&&(typeof o!="boolean"||o===a.options)){var s=a.target.__events__,l=s[a.eventName],u=l?l[this._id]:null;u&&(u.length===1||!r?(l.count-=u.length,delete s[a.eventName][this._id]):(l.count--,u.splice(u.indexOf(a),1)),l.count||delete s[a.eventName]),a.elementCallback&&(a.target.removeEventListener?a.target.removeEventListener(a.eventName,a.elementCallback,a.options):a.target.detachEvent&&a.target.detachEvent("on"+a.eventName,a.elementCallback)),this._eventRecords.splice(i--,1)}}},e.prototype.raise=function(t,n,r){return e.raise(this._parent,t,n,r)},e.prototype.declare=function(t){var n=this._parent.__declaredEvents=this._parent.__declaredEvents||{};if(typeof t=="string")n[t]=!0;else for(var r=0;r<t.length;r++)n[t[r]]=!0},e._uniqueId=0,e}();function ss(e){if(!(typeof document>"u")){var t=e;return t&&t.ownerDocument?t.ownerDocument:document}}var Z5;Fo({overflow:"hidden !important"});function MG(){if(Z5===void 0){var e=document.createElement("div");e.style.setProperty("width","100px"),e.style.setProperty("height","100px"),e.style.setProperty("overflow","scroll"),e.style.setProperty("position","absolute"),e.style.setProperty("top","-9999px"),document.body.appendChild(e),Z5=e.offsetWidth-e.clientWidth,document.body.removeChild(e)}return Z5}var LG=void 0;function xF(e){console&&console.warn&&console.warn(e)}function CF(e,t,n,r,o){if(o===!0&&!1)for(var i,a;i<a.length;i++)var s}(function(e){yi(t,e);function t(n,r){var o=e.call(this,n,r)||this;return jG(o,t.prototype,["componentDidMount","shouldComponentUpdate","getSnapshotBeforeUpdate","render","componentDidUpdate","componentWillUnmount"]),o}return t.prototype.componentDidUpdate=function(n,r){this._updateComponentRef(n,this.props)},t.prototype.componentDidMount=function(){this._setComponentRef(this.props.componentRef,this)},t.prototype.componentWillUnmount=function(){if(this._setComponentRef(this.props.componentRef,null),this.__disposables){for(var n=0,r=this._disposables.length;n<r;n++){var o=this.__disposables[n];o.dispose&&o.dispose()}this.__disposables=null}},Object.defineProperty(t.prototype,"className",{get:function(){if(!this.__className){var n=/function (.{1,})\(/,r=n.exec(this.constructor.toString());this.__className=r&&r.length>1?r[1]:""}return this.__className},enumerable:!1,configurable:!0}),Object.defineProperty(t.prototype,"_disposables",{get:function(){return this.__disposables||(this.__disposables=[]),this.__disposables},enumerable:!1,configurable:!0}),Object.defineProperty(t.prototype,"_async",{get:function(){return this.__async||(this.__async=new _b(this),this._disposables.push(this.__async)),this.__async},enumerable:!1,configurable:!0}),Object.defineProperty(t.prototype,"_events",{get:function(){return this.__events||(this.__events=new Ws(this),this._disposables.push(this.__events)),this.__events},enumerable:!1,configurable:!0}),t.prototype._resolveRef=function(n){var r=this;return this.__resolves||(this.__resolves={}),this.__resolves[n]||(this.__resolves[n]=function(o){return r[n]=o}),this.__resolves[n]},t.prototype._updateComponentRef=function(n,r){r===void 0&&(r={}),n&&r&&n.componentRef!==r.componentRef&&(this._setComponentRef(n.componentRef,null),this._setComponentRef(r.componentRef,this))},t.prototype._warnDeprecations=function(n){this.className,this.props},t.prototype._warnMutuallyExclusive=function(n){this.className,this.props},t.prototype._warnConditionallyRequiredProps=function(n,r,o){CF(this.className,this.props,n,r,o)},t.prototype._setComponentRef=function(n,r){!this._skipComponentRefResolution&&n&&(typeof n=="function"&&n(r),typeof n=="object"&&(n.current=r))},t})(T.Component);function jG(e,t,n){for(var r=0,o=n.length;r<o;r++)zG(e,t,n[r])}function zG(e,t,n){var r=e[n],o=t[n];(r||o)&&(e[n]=function(){for(var i=[],a=0;a<arguments.length;a++)i[a]=arguments[a];var s;return o&&(s=o.apply(this,i)),r!==o&&(s=r.apply(this,i)),s})}function AF(){return null}var J5="__globalSettings__",w4="__callbacks__",HG=0,NF=function(){function e(){}return e.getValue=function(t,n){var r=vE();return r[t]===void 0&&(r[t]=typeof n=="function"?n():n),r[t]},e.setValue=function(t,n){var r=vE(),o=r[w4],i=r[t];if(n!==i){r[t]=n;var a={oldValue:i,value:n,key:t};for(var s in o)o.hasOwnProperty(s)&&o[s](a)}return n},e.addChangeListener=function(t){var n=t.__id__,r=tx();n||(n=t.__id__=String(HG++)),r[n]=t},e.removeChangeListener=function(t){var n=tx();delete n[t.__id__]},e}();function vE(){var e,t=ur(),n=t||{};return n[J5]||(n[J5]=(e={},e[w4]={},e)),n[J5]}function tx(){var e=vE();return e[w4]}var qt={backspace:8,tab:9,enter:13,shift:16,ctrl:17,alt:18,pauseBreak:19,capslock:20,escape:27,space:32,pageUp:33,pageDown:34,end:35,home:36,left:37,up:38,right:39,down:40,insert:45,del:46,zero:48,one:49,two:50,three:51,four:52,five:53,six:54,seven:55,eight:56,nine:57,colon:58,a:65,b:66,c:67,d:68,e:69,f:70,g:71,h:72,i:73,j:74,k:75,l:76,m:77,n:78,o:79,p:80,q:81,r:82,s:83,t:84,u:85,v:86,w:87,x:88,y:89,z:90,leftWindow:91,rightWindow:92,select:93,zero_numpad:96,one_numpad:97,two_numpad:98,three_numpad:99,four_numpad:100,five_numpad:101,six_numpad:102,seven_numpad:103,eight_numpad:104,nine_numpad:105,multiply:106,add:107,subtract:109,decimalPoint:110,divide:111,f1:112,f2:113,f3:114,f4:115,f5:116,f6:117,f7:118,f8:119,f9:120,f10:121,f11:122,f12:123,numlock:144,scrollLock:145,semicolon:186,equalSign:187,comma:188,dash:189,period:190,forwardSlash:191,graveAccent:192,openBracket:219,backSlash:220,closeBracket:221,singleQuote:222},ts=function(){function e(t,n,r,o){t===void 0&&(t=0),n===void 0&&(n=0),r===void 0&&(r=0),o===void 0&&(o=0),this.top=r,this.bottom=o,this.left=t,this.right=n}return Object.defineProperty(e.prototype,"width",{get:function(){return this.right-this.left},enumerable:!1,configurable:!0}),Object.defineProperty(e.prototype,"height",{get:function(){return this.bottom-this.top},enumerable:!1,configurable:!0}),e.prototype.equals=function(t){return parseFloat(this.top.toFixed(4))===parseFloat(t.top.toFixed(4))&&parseFloat(this.bottom.toFixed(4))===parseFloat(t.bottom.toFixed(4))&&parseFloat(this.left.toFixed(4))===parseFloat(t.left.toFixed(4))&&parseFloat(this.right.toFixed(4))===parseFloat(t.right.toFixed(4))},e}();function UG(e){for(var t=[],n=1;n<arguments.length;n++)t[n-1]=arguments[n];return t.length<2?t[0]:function(){for(var r=[],o=0;o<arguments.length;o++)r[o]=arguments[o];t.forEach(function(i){return i&&i.apply(e,r)})}}function $0(){for(var e=[],t=0;t<arguments.length;t++)e[t]=arguments[t];var n=e.filter(function(r){return r}).join(" ").trim();return n===""?void 0:n}function qG(e,t,n){var r=e.slice();return r.splice(t,0,n),r}function $G(e,t){if(e.length!==t.length)return!1;for(var n=0;n<e.length;n++)if(e[n]!==t[n])return!1;return!0}function FF(e){var t=null;try{var n=ur();t=n?n.sessionStorage.getItem(e):null}catch{}return t}function WG(e,t){var n;try{(n=ur())===null||n===void 0||n.sessionStorage.setItem(e,t)}catch{}}var IF="isRTL",Hs;function ls(e){if(e===void 0&&(e={}),e.rtl!==void 0)return e.rtl;if(Hs===void 0){var t=FF(IF);t!==null&&(Hs=t==="1",GG(Hs));var n=ss();Hs===void 0&&n&&(Hs=(n.body&&n.body.getAttribute("dir")||n.documentElement.getAttribute("dir"))==="rtl",gF(Hs))}return!!Hs}function GG(e,t){t===void 0&&(t=!1);var n=ss();n&&n.documentElement.setAttribute("dir",e?"rtl":"ltr"),t&&WG(IF,e?"1":"0"),Hs=e,gF(Hs)}function KG(e){return e&&!!e._virtual}function VG(e){var t;return e&&KG(e)&&(t=e._virtual.parent),t}function BF(e,t){return t===void 0&&(t=!0),e&&(t&&VG(e)||e.parentNode&&e.parentNode)}function bE(e,t,n){n===void 0&&(n=!0);var r=!1;if(e&&t)if(n)if(e===t)r=!0;else for(r=!1;t;){var o=BF(t);if(o===e){r=!0;break}t=o}else e.contains&&(r=e.contains(t));return r}function k4(e,t){return!e||e===document.body?null:t(e)?e:k4(BF(e),t)}function YG(e,t){var n=k4(e,function(r){return r.hasAttribute(t)});return n&&n.getAttribute(t)}var yE="data-portal-element";function QG(e){e.setAttribute(yE,"true")}function XG(e,t){var n=k4(e,function(r){return t===r||r.hasAttribute(yE)});return n!==null&&n.hasAttribute(yE)}function ZG(e,t){var n=e,r=t;n._virtual||(n._virtual={children:[]});var o=n._virtual.parent;if(o&&o!==t){var i=o._virtual.children.indexOf(n);i>-1&&o._virtual.children.splice(i,1)}n._virtual.parent=r||void 0,r&&(r._virtual||(r._virtual={children:[]}),r._virtual.children.push(n))}var JG="data-is-focusable",eK="data-is-visible",tK="data-focuszone-id",nK="data-is-sub-focuszone";function rK(e,t,n){return xh(e,t,!0,!1,!1,n)}function oK(e,t,n){return jf(e,t,!0,!1,!0,n)}function iK(e){var t=xh(e,e,!0,!1,!1,!0);return t?(lK(t),!0):!1}function jf(e,t,n,r,o,i,a,s){if(!t||!a&&t===e)return null;var l=RF(t);if(o&&l&&(i||!(OF(t)||DF(t)))){var u=jf(e,t.lastElementChild,!0,!0,!0,i,a,s);if(u){if(s&&EE(u,!0)||!s)return u;var d=jf(e,u.previousElementSibling,!0,!0,!0,i,a,s);if(d)return d;for(var h=u.parentElement;h&&h!==t;){var p=jf(e,h.previousElementSibling,!0,!0,!0,i,a,s);if(p)return p;h=h.parentElement}}}if(n&&l&&EE(t,s))return t;var m=jf(e,t.previousElementSibling,!0,!0,!0,i,a,s);return m||(r?null:jf(e,t.parentElement,!0,!1,!1,i,a,s))}function xh(e,t,n,r,o,i,a,s){if(!t||t===e&&o&&!a)return null;var l=RF(t);if(n&&l&&EE(t,s))return t;if(!o&&l&&(i||!(OF(t)||DF(t)))){var u=xh(e,t.firstElementChild,!0,!0,!1,i,a,s);if(u)return u}if(t===e)return null;var d=xh(e,t.nextElementSibling,!0,!0,!1,i,a,s);return d||(r?null:xh(e,t.parentElement,!1,!1,!0,i,a,s))}function RF(e){if(!e||!e.getAttribute)return!1;var t=e.getAttribute(eK);return t!=null?t==="true":e.offsetHeight!==0||e.offsetParent!==null||e.isVisible===!0}function EE(e,t){if(!e||e.disabled)return!1;var n=0,r=null;e&&e.getAttribute&&(r=e.getAttribute("tabIndex"),r&&(n=parseInt(r,10)));var o=e.getAttribute?e.getAttribute(JG):null,i=r!==null&&n>=0,a=!!e&&o!=="false"&&(e.tagName==="A"||e.tagName==="BUTTON"||e.tagName==="INPUT"||e.tagName==="TEXTAREA"||e.tagName==="SELECT"||o==="true"||i);return t?n!==-1&&a:a}function OF(e){return!!(e&&e.getAttribute&&e.getAttribute(tK))}function DF(e){return!!(e&&e.getAttribute&&e.getAttribute(nK)==="true")}function aK(e){var t=ss(e),n=t&&t.activeElement;return!!(n&&bE(e,n))}function sK(e,t){return YG(e,t)!=="true"}var bf=void 0;function lK(e){if(e){if(bf){bf=e;return}bf=e;var t=ur(e);t&&t.requestAnimationFrame(function(){bf&&bf.focus(),bf=void 0})}}function zf(e,t,n,r){return e.addEventListener(t,n,r),function(){return e.removeEventListener(t,n,r)}}var uK=50,cK=5,yg=0,e2=Qi.getInstance();e2&&e2.onReset&&e2.onReset(function(){return yg++});var ym="__retval__";function ml(e){e===void 0&&(e={});var t=new Map,n=0,r=0,o=yg,i=function(a,s){var l;if(s===void 0&&(s={}),e.useStaticStyles&&typeof a=="function"&&a.__noStyleOverride__)return a(s);r++;var u=t,d=s.theme,h=d&&d.rtl!==void 0?d.rtl:ls(),p=e.disableCaching;if(o!==yg&&(o=yg,t=new Map,n=0),e.disableCaching||(u=nx(t,a),u=nx(u,s)),(p||!u[ym])&&(a===void 0?u[ym]={}:u[ym]=wF([typeof a=="function"?a(s):a],{rtl:!!h,specificityMultiplier:e.useStaticStyles?cK:void 0}),p||n++),n>(e.cacheSize||uK)){var m=ur();!((l=m==null?void 0:m.FabricConfig)===null||l===void 0)&&l.enableClassNameCacheFullWarning&&(console.warn("Styles are being recalculated too frequently. Cache miss rate is "+n+"/"+r+"."),console.trace()),t.clear(),n=0,e.disableCaching=!0}return u[ym]};return i}function t2(e,t){return t=fK(t),e.has(t)||e.set(t,new Map),e.get(t)}function nx(e,t){if(typeof t=="function"){var n=t.__cachedInputs__;if(n)for(var r=0,o=t.__cachedInputs__;r<o.length;r++){var i=o[r];e=t2(e,i)}else e=t2(e,t)}else if(typeof t=="object")for(var a in t)t.hasOwnProperty(a)&&(e=t2(e,t[a]));return e}function fK(e){switch(e){case void 0:return"__undefined__";case null:return"__null__";default:return e}}var rx=!1,Eg=0,dK={empty:!0},n2={},r0=typeof WeakMap>"u"?null:WeakMap;function hK(){Eg++}function Cr(e,t,n){if(t===void 0&&(t=100),n===void 0&&(n=!1),!r0)return e;if(!rx){var r=Qi.getInstance();r&&r.onReset&&Qi.getInstance().onReset(hK),rx=!0}var o,i=0,a=Eg;return function(){for(var l=[],u=0;u<arguments.length;u++)l[u]=arguments[u];var d=o;(o===void 0||a!==Eg||t>0&&i>t)&&(o=ox(),i=0,a=Eg),d=o;for(var h=0;h<l.length;h++){var p=pK(l[h]);d.map.has(p)||d.map.set(p,ox()),d=d.map.get(p)}return d.hasOwnProperty("value")||(d.value=e.apply(void 0,l),i++),n&&(d.value===null||d.value===void 0)&&(d.value=e.apply(void 0,l)),d.value}}function Od(e){if(!r0)return e;var t=new r0;function n(r){if(!r||typeof r!="function"&&typeof r!="object")return e(r);if(t.has(r))return t.get(r);var o=e(r);return t.set(r,o),o}return n}function pK(e){if(e){if(typeof e=="object"||typeof e=="function")return e;n2[e]||(n2[e]={val:e})}else return dK;return n2[e]}function ox(){return{map:r0?new r0:null}}function mK(e){var t=e,n=Od(function(r){if(e===r)throw new Error("Attempted to compose a component with itself.");var o=r,i=Od(function(s){var l=function(u){return T.createElement(o,oe({},u,{defaultRender:s}))};return l}),a=function(s){var l=s.defaultRender;return T.createElement(t,oe({},s,{defaultRender:l?i(l):o}))};return a});return n}var gK=Od(mK);function PF(e,t){return gK(e)(t)}function Ys(){for(var e=[],t=0;t<arguments.length;t++)e[t]=arguments[t];for(var n=[],r=0,o=e;r<o.length;r++){var i=o[r];if(i)if(typeof i=="string")n.push(i);else if(i.hasOwnProperty("toString")&&typeof i.toString=="function")n.push(i.toString());else for(var a in i)i[a]&&n.push(a)}return n.join(" ")}var vK="customizations",bK={settings:{},scopedSettings:{},inCustomizerContext:!1},Ol=NF.getValue(vK,{settings:{},scopedSettings:{},inCustomizerContext:!1}),Em=[],$i=function(){function e(){}return e.reset=function(){Ol.settings={},Ol.scopedSettings={}},e.applySettings=function(t){Ol.settings=oe(oe({},Ol.settings),t),e._raiseChange()},e.applyScopedSettings=function(t,n){Ol.scopedSettings[t]=oe(oe({},Ol.scopedSettings[t]),n),e._raiseChange()},e.getSettings=function(t,n,r){r===void 0&&(r=bK);for(var o={},i=n&&r.scopedSettings[n]||{},a=n&&Ol.scopedSettings[n]||{},s=0,l=t;s<l.length;s++){var u=l[s];o[u]=i[u]||r.settings[u]||a[u]||Ol.settings[u]}return o},e.applyBatchedUpdates=function(t,n){e._suppressUpdates=!0;try{t()}catch{}e._suppressUpdates=!1,n||e._raiseChange()},e.observe=function(t){Em.push(t)},e.unobserve=function(t){Em=Em.filter(function(n){return n!==t})},e._raiseChange=function(){e._suppressUpdates||Em.forEach(function(t){return t()})},e}(),o0=T.createContext({customizations:{inCustomizerContext:!1,settings:{},scopedSettings:{}}});function yK(e,t){e===void 0&&(e={});var n=MF(t)?t:_K(t);return n(e)}function EK(e,t){e===void 0&&(e={});var n=MF(t)?t:TK(t);return n(e)}function MF(e){return typeof e=="function"}function _K(e){return function(t){return e?oe(oe({},t),e):t}}function TK(e){return e===void 0&&(e={}),function(t){var n=oe({},t);for(var r in e)e.hasOwnProperty(r)&&(n[r]=oe(oe({},t[r]),e[r]));return n}}function wK(e,t){var n=(t||{}).customizations,r=n===void 0?{settings:{},scopedSettings:{}}:n;return{customizations:{settings:yK(r.settings,e.settings),scopedSettings:EK(r.scopedSettings,e.scopedSettings),inCustomizerContext:!0}}}var kK=function(e){yi(t,e);function t(){var n=e!==null&&e.apply(this,arguments)||this;return n._onCustomizationChange=function(){return n.forceUpdate()},n}return t.prototype.componentDidMount=function(){$i.observe(this._onCustomizationChange)},t.prototype.componentWillUnmount=function(){$i.unobserve(this._onCustomizationChange)},t.prototype.render=function(){var n=this,r=this.props.contextTransform;return T.createElement(o0.Consumer,null,function(o){var i=wK(n.props,o);return r&&(i=r(i)),T.createElement(o0.Provider,{value:i},n.props.children)})},t}(T.Component);function SK(e,t){for(var n in e)e.hasOwnProperty(n)&&(t[n]=e[n]);return t}function xK(e,t,n){return function(o){var i,a=(i=function(s){yi(l,s);function l(u){var d=s.call(this,u)||this;return d._styleCache={},d._onSettingChanged=d._onSettingChanged.bind(d),d}return l.prototype.componentDidMount=function(){$i.observe(this._onSettingChanged)},l.prototype.componentWillUnmount=function(){$i.unobserve(this._onSettingChanged)},l.prototype.render=function(){var u=this;return T.createElement(o0.Consumer,null,function(d){var h=$i.getSettings(t,e,d.customizations),p=u.props;if(h.styles&&typeof h.styles=="function"&&(h.styles=h.styles(oe(oe({},h),p))),n&&h.styles){if(u._styleCache.default!==h.styles||u._styleCache.component!==p.styles){var m=jc(h.styles,p.styles);u._styleCache.default=h.styles,u._styleCache.component=p.styles,u._styleCache.merged=m}return T.createElement(o,oe({},h,p,{styles:u._styleCache.merged}))}return T.createElement(o,oe({},h,p))})},l.prototype._onSettingChanged=function(){this.forceUpdate()},l}(T.Component),i.displayName="Customized"+e,i);return SK(o,a)}}function CK(e,t){var n=AK(),r=T.useContext(o0).customizations,o=r.inCustomizerContext;return T.useEffect(function(){return o||$i.observe(n),function(){o||$i.unobserve(n)}},[o]),$i.getSettings(e,t,r)}function AK(){var e=T.useState(0),t=e[1];return function(){return t(function(n){return++n})}}function NK(e,t){for(var n in t)t.hasOwnProperty(n)&&(e[n]=UG(e,e[n],t[n]))}var ov="__currentId__",FK="id__",iv=ur()||{};iv[ov]===void 0&&(iv[ov]=0);var ix=!1;function cu(e){if(!ix){var t=Qi.getInstance();t&&t.onReset&&t.onReset(IK),ix=!0}var n=iv[ov]++;return(e===void 0?FK:e)+n}function IK(e){e===void 0&&(e=0),iv[ov]=e}var Xn=function(){for(var e=[],t=0;t<arguments.length;t++)e[t]=arguments[t];for(var n={},r=0,o=e;r<o.length;r++)for(var i=o[r],a=Array.isArray(i)?i:Object.keys(i),s=0,l=a;s<l.length;s++){var u=l[s];n[u]=1}return n},BK=Xn(["onCopy","onCut","onPaste","onCompositionEnd","onCompositionStart","onCompositionUpdate","onFocus","onFocusCapture","onBlur","onBlurCapture","onChange","onInput","onSubmit","onLoad","onError","onKeyDown","onKeyDownCapture","onKeyPress","onKeyUp","onAbort","onCanPlay","onCanPlayThrough","onDurationChange","onEmptied","onEncrypted","onEnded","onLoadedData","onLoadedMetadata","onLoadStart","onPause","onPlay","onPlaying","onProgress","onRateChange","onSeeked","onSeeking","onStalled","onSuspend","onTimeUpdate","onVolumeChange","onWaiting","onClick","onClickCapture","onContextMenu","onDoubleClick","onDrag","onDragEnd","onDragEnter","onDragExit","onDragLeave","onDragOver","onDragStart","onDrop","onMouseDown","onMouseDownCapture","onMouseEnter","onMouseLeave","onMouseMove","onMouseOut","onMouseOver","onMouseUp","onMouseUpCapture","onSelect","onTouchCancel","onTouchEnd","onTouchMove","onTouchStart","onScroll","onWheel","onPointerCancel","onPointerDown","onPointerEnter","onPointerLeave","onPointerMove","onPointerOut","onPointerOver","onPointerUp","onGotPointerCapture","onLostPointerCapture"]),RK=Xn(["accessKey","children","className","contentEditable","dir","draggable","hidden","htmlFor","id","lang","ref","role","style","tabIndex","title","translate","spellCheck","name"]),fr=Xn(RK,BK);Xn(fr,["form"]);var OK=Xn(fr,["height","loop","muted","preload","src","width"]);Xn(OK,["poster"]);Xn(fr,["start"]);Xn(fr,["value"]);var LF=Xn(fr,["download","href","hrefLang","media","rel","target","type"]),Oc=Xn(fr,["autoFocus","disabled","form","formAction","formEncType","formMethod","formNoValidate","formTarget","type","value"]);Xn(Oc,["accept","alt","autoCapitalize","autoComplete","checked","dirname","form","height","inputMode","list","max","maxLength","min","minLength","multiple","pattern","placeholder","readOnly","required","src","step","size","type","value","width"]);Xn(Oc,["autoCapitalize","cols","dirname","form","maxLength","minLength","placeholder","readOnly","required","rows","wrap"]);Xn(Oc,["form","multiple","required"]);Xn(fr,["selected","value"]);Xn(fr,["cellPadding","cellSpacing"]);Xn(fr,["rowSpan","scope"]);Xn(fr,["colSpan","headers","rowSpan","scope"]);Xn(fr,["span"]);Xn(fr,["span"]);Xn(fr,["acceptCharset","action","encType","encType","method","noValidate","target"]);Xn(fr,["allow","allowFullScreen","allowPaymentRequest","allowTransparency","csp","height","importance","referrerPolicy","sandbox","src","srcDoc","width"]);var DK=Xn(fr,["alt","crossOrigin","height","src","srcSet","useMap","width"]),W0=fr;function Zr(e,t,n){for(var r=Array.isArray(t),o={},i=Object.keys(e),a=0,s=i;a<s.length;a++){var l=s[a],u=!r&&t[l]||r&&t.indexOf(l)>=0||l.indexOf("data-")===0||l.indexOf("aria-")===0;u&&(!n||(n==null?void 0:n.indexOf(l))===-1)&&(o[l]=e[l])}return o}function Tb(e){NK(e,{componentDidMount:PK,componentDidUpdate:MK,componentWillUnmount:LK})}function PK(){av(this.props.componentRef,this)}function MK(e){e.componentRef!==this.props.componentRef&&(av(e.componentRef,null),av(this.props.componentRef,this))}function LK(){av(this.props.componentRef,null)}function av(e,t){e&&(typeof e=="object"?e.current=t:typeof e=="function"&&e(t))}var ca,jK=(ca={},ca[qt.up]=1,ca[qt.down]=1,ca[qt.left]=1,ca[qt.right]=1,ca[qt.home]=1,ca[qt.end]=1,ca[qt.tab]=1,ca[qt.pageUp]=1,ca[qt.pageDown]=1,ca);function zK(e){return!!jK[e]}var Go="ms-Fabric--isFocusVisible",ax="ms-Fabric--isFocusHidden";function dd(e,t){var n=t?ur(t):ur();if(n){var r=n.document.body.classList;r.add(e?Go:ax),r.remove(e?ax:Go)}}var sx=new WeakMap;function lx(e,t){var n,r=sx.get(e);return r?n=r+t:n=1,sx.set(e,n),n}function jF(e){T.useEffect(function(){var t,n=ur(e==null?void 0:e.current);if(!(!n||((t=n.FabricConfig)===null||t===void 0?void 0:t.disableFocusRects)===!0)){var r=lx(n,1);return r<=1&&(n.addEventListener("mousedown",ux,!0),n.addEventListener("pointerdown",cx,!0),n.addEventListener("keydown",fx,!0)),function(){var o;!n||((o=n.FabricConfig)===null||o===void 0?void 0:o.disableFocusRects)===!0||(r=lx(n,-1),r===0&&(n.removeEventListener("mousedown",ux,!0),n.removeEventListener("pointerdown",cx,!0),n.removeEventListener("keydown",fx,!0)))}}},[e])}var HK=function(e){return jF(e.rootRef),null};function ux(e){dd(!1,e.target)}function cx(e){e.pointerType!=="mouse"&&dd(!1,e.target)}function fx(e){zK(e.which)&&dd(!0,e.target)}function UK(e){var t=null;try{var n=ur();t=n?n.localStorage.getItem(e):null}catch{}return t}var tc,dx="language";function qK(e){if(e===void 0&&(e="sessionStorage"),tc===void 0){var t=ss(),n=e==="localStorage"?UK(dx):e==="sessionStorage"?FF(dx):void 0;n&&(tc=n),tc===void 0&&t&&(tc=t.documentElement.getAttribute("lang")),tc===void 0&&(tc="en")}return tc}function hx(e){for(var t=[],n=1;n<arguments.length;n++)t[n-1]=arguments[n];for(var r=0,o=t;r<o.length;r++){var i=o[r];zF(e||{},i)}return e}function zF(e,t,n){n===void 0&&(n=[]),n.push(t);for(var r in t)if(t.hasOwnProperty(r)&&r!=="__proto__"&&r!=="constructor"&&r!=="prototype"){var o=t[r];if(typeof o=="object"&&o!==null&&!Array.isArray(o)){var i=n.indexOf(o)>-1;e[r]=i?o:zF(e[r]||{},o,n)}else e[r]=o}return n.pop(),e}var px=function(){return!window||!window.navigator||!window.navigator.userAgent?!1:/iPad|iPhone|iPod/i.test(window.navigator.userAgent)},$K=["TEMPLATE","STYLE","SCRIPT"];function WK(e){var t=ss(e);if(!t)return function(){};for(var n=[];e!==t.body&&e.parentElement;){for(var r=0,o=e.parentElement.children;r<o.length;r++){var i=o[r],a=i.getAttribute("aria-hidden");i!==e&&(a==null?void 0:a.toLowerCase())!=="true"&&$K.indexOf(i.tagName)===-1&&n.push([i,a])}e=e.parentElement}return n.forEach(function(s){var l=s[0];l.setAttribute("aria-hidden","true")}),function(){GK(n),n=[]}}function GK(e){e.forEach(function(t){var n=t[0],r=t[1];r?n.setAttribute("aria-hidden",r):n.removeAttribute("aria-hidden")})}var r2;function mx(e){var t;if(typeof r2>"u"||e){var n=ur(),r=(t=n==null?void 0:n.navigator)===null||t===void 0?void 0:t.userAgent;r2=!!r&&r.indexOf("Macintosh")!==-1}return!!r2}function KK(e){var t=Od(function(n){var r=Od(function(o){return function(i){return n(i,o)}});return function(o,i){return e(o,i?r(i):n)}});return t}var VK=Od(KK);function HF(e,t){return VK(e)(t)}var YK=["theme","styles"];function gl(e,t,n,r,o){r=r||{scope:"",fields:void 0};var i=r.scope,a=r.fields,s=a===void 0?YK:a,l=T.forwardRef(function(d,h){var p=T.useRef(),m=CK(s,i),v=m.styles;m.dir;var _=pl(m,["styles","dir"]),b=n?n(d):void 0,E=p.current&&p.current.__cachedInputs__||[],w=d.styles;if(!p.current||v!==E[1]||w!==E[2]){var k=function(y){return kF(y,t,v,w)};k.__cachedInputs__=[t,v,w],k.__noStyleOverride__=!v&&!w,p.current=k}return T.createElement(e,oe({ref:h},_,b,d,{styles:p.current}))});l.displayName="Styled"+(e.displayName||e.name);var u=o?T.memo(l):l;return l.displayName&&(u.displayName=l.displayName),u}function S4(e,t){for(var n=oe({},t),r=0,o=Object.keys(e);r<o.length;r++){var i=o[r];n[i]===void 0&&(n[i]=e[i])}return n}var QK=function(e){return function(t){for(var n=0,r=e.refs;n<r.length;n++){var o=r[n];typeof o=="function"?o(t):o&&(o.current=t)}}},XK=function(e){var t={refs:[]};return function(){for(var n=[],r=0;r<arguments.length;r++)n[r]=arguments[r];return(!t.resolver||!$G(t.refs,n))&&(t.resolver=QK(t)),t.refs=n,t.resolver}},i0=T.useLayoutEffect,ZK="icons",mi=NF.getValue(ZK,{__options:{disableWarnings:!1,warnOnMissingIcons:!0},__remapped:{}}),o2=Qi.getInstance();o2&&o2.onReset&&o2.onReset(function(){for(var e in mi)mi.hasOwnProperty(e)&&mi[e].subset&&(mi[e].subset.className=void 0)});var sv=function(e){return e.toLowerCase()};function Ar(e,t){var n=oe(oe({},e),{isRegistered:!1,className:void 0}),r=e.icons;t=t?oe(oe({},mi.__options),t):mi.__options;for(var o in r)if(r.hasOwnProperty(o)){var i=r[o],a=sv(o);mi[a]?eV(o):mi[a]={code:i,subset:n}}}function nc(e,t){mi.__remapped[sv(e)]=sv(t)}function JK(e){var t=void 0,n=mi.__options;if(e=e?sv(e):"",e=mi.__remapped[e]||e,e)if(t=mi[e],t){var r=t.subset;r&&r.fontFace&&(r.isRegistered||(SF(r.fontFace),r.isRegistered=!0),r.className||(r.className=Fo(r.style,{fontFamily:r.fontFace.fontFamily,fontWeight:r.fontFace.fontWeight||"normal",fontStyle:r.fontFace.fontStyle||"normal"})))}else!n.disableWarnings&&n.warnOnMissingIcons&&xF('The icon "'+e+'" was used but not registered. See https://github.com/microsoft/fluentui/wiki/Using-icons for more information.');return t}var z1=[],i2=void 0;function eV(e){var t=mi.__options,n=2e3,r=10;t.disableWarnings||(z1.push(e),i2===void 0&&(i2=setTimeout(function(){xF(`Some icons were re-registered. Applications should only call registerIcons for any given icon once. Redefining what an icon is may have unintended consequences. Duplicates include: 
+***************************************************************************** */var mE=function(e,t){return mE=Object.setPrototypeOf||{__proto__:[]}instanceof Array&&function(n,r){n.__proto__=r}||function(n,r){for(var o in r)Object.prototype.hasOwnProperty.call(r,o)&&(n[o]=r[o])},mE(e,t)};function yi(e,t){if(typeof t!="function"&&t!==null)throw new TypeError("Class extends value "+String(t)+" is not a constructor or null");mE(e,t);function n(){this.constructor=e}e.prototype=t===null?Object.create(t):(n.prototype=t.prototype,new n)}var oe=function(){return oe=Object.assign||function(t){for(var n,r=1,o=arguments.length;r<o;r++){n=arguments[r];for(var i in n)Object.prototype.hasOwnProperty.call(n,i)&&(t[i]=n[i])}return t},oe.apply(this,arguments)};function pl(e,t){var n={};for(var r in e)Object.prototype.hasOwnProperty.call(e,r)&&t.indexOf(r)<0&&(n[r]=e[r]);if(e!=null&&typeof Object.getOwnPropertySymbols=="function")for(var o=0,r=Object.getOwnPropertySymbols(e);o<r.length;o++)t.indexOf(r[o])<0&&Object.prototype.propertyIsEnumerable.call(e,r[o])&&(n[r[o]]=e[r[o]]);return n}function vG(e,t,n,r){var o=arguments.length,i=o<3?t:r===null?r=Object.getOwnPropertyDescriptor(t,n):r,a;if(typeof Reflect=="object"&&typeof Reflect.decorate=="function")i=Reflect.decorate(e,t,n,r);else for(var s=e.length-1;s>=0;s--)(a=e[s])&&(i=(o<3?a(i):o>3?a(t,n,i):a(t,n))||i);return o>3&&i&&Object.defineProperty(t,n,i),i}function Yi(e,t,n){if(n||arguments.length===2)for(var r=0,o=t.length,i;r<o;r++)(i||!(r in t))&&(i||(i=Array.prototype.slice.call(t,0,r)),i[r]=t[r]);return e.concat(i||t)}var K8={},bg=void 0;try{bg=window}catch{}function yb(e,t){if(typeof bg<"u"){var n=bg.__packages__=bg.__packages__||{};if(!n[e]||!K8[e]){K8[e]=t;var r=n[e]=n[e]||[];r.push(t)}}}yb("@fluentui/set-version","6.0.0");var j1={none:0,insertNode:1,appendChild:2},V8="__stylesheet__",bG=typeof navigator<"u"&&/rv:11.0/.test(navigator.userAgent),Lf={};try{Lf=window||{}}catch{}var vf,Qi=function(){function e(t,n){var r,o,i,a,s,l;this._rules=[],this._preservedRules=[],this._counter=0,this._keyToClassName={},this._onInsertRuleCallbacks=[],this._onResetCallbacks=[],this._classNameToArgs={},this._config=oe({injectionMode:typeof document>"u"?j1.none:j1.insertNode,defaultPrefix:"css",namespace:void 0,cspSettings:void 0},t),this._classNameToArgs=(r=n==null?void 0:n.classNameToArgs)!==null&&r!==void 0?r:this._classNameToArgs,this._counter=(o=n==null?void 0:n.counter)!==null&&o!==void 0?o:this._counter,this._keyToClassName=(a=(i=this._config.classNameCache)!==null&&i!==void 0?i:n==null?void 0:n.keyToClassName)!==null&&a!==void 0?a:this._keyToClassName,this._preservedRules=(s=n==null?void 0:n.preservedRules)!==null&&s!==void 0?s:this._preservedRules,this._rules=(l=n==null?void 0:n.rules)!==null&&l!==void 0?l:this._rules}return e.getInstance=function(){if(vf=Lf[V8],!vf||vf._lastStyleElement&&vf._lastStyleElement.ownerDocument!==document){var t=(Lf==null?void 0:Lf.FabricConfig)||{},n=new e(t.mergeStyles,t.serializedStylesheet);vf=n,Lf[V8]=n}return vf},e.prototype.serialize=function(){return JSON.stringify({classNameToArgs:this._classNameToArgs,counter:this._counter,keyToClassName:this._keyToClassName,preservedRules:this._preservedRules,rules:this._rules})},e.prototype.setConfig=function(t){this._config=oe(oe({},this._config),t)},e.prototype.onReset=function(t){var n=this;return this._onResetCallbacks.push(t),function(){n._onResetCallbacks=n._onResetCallbacks.filter(function(r){return r!==t})}},e.prototype.onInsertRule=function(t){var n=this;return this._onInsertRuleCallbacks.push(t),function(){n._onInsertRuleCallbacks=n._onInsertRuleCallbacks.filter(function(r){return r!==t})}},e.prototype.getClassName=function(t){var n=this._config.namespace,r=t||this._config.defaultPrefix;return(n?n+"-":"")+r+"-"+this._counter++},e.prototype.cacheClassName=function(t,n,r,o){this._keyToClassName[n]=t,this._classNameToArgs[t]={args:r,rules:o}},e.prototype.classNameFromKey=function(t){return this._keyToClassName[t]},e.prototype.getClassNameCache=function(){return this._keyToClassName},e.prototype.argsFromClassName=function(t){var n=this._classNameToArgs[t];return n&&n.args},e.prototype.insertedRulesFromClassName=function(t){var n=this._classNameToArgs[t];return n&&n.rules},e.prototype.insertRule=function(t,n){var r=this._config.injectionMode,o=r!==j1.none?this._getStyleElement():void 0;if(n&&this._preservedRules.push(t),o)switch(r){case j1.insertNode:var i=o.sheet;try{i.insertRule(t,i.cssRules.length)}catch{}break;case j1.appendChild:o.appendChild(document.createTextNode(t));break}else this._rules.push(t);this._config.onInsertRule&&this._config.onInsertRule(t),this._onInsertRuleCallbacks.forEach(function(a){return a()})},e.prototype.getRules=function(t){return(t?this._preservedRules.join(""):"")+this._rules.join("")},e.prototype.reset=function(){this._rules=[],this._counter=0,this._classNameToArgs={},this._keyToClassName={},this._onResetCallbacks.forEach(function(t){return t()})},e.prototype.resetKeys=function(){this._keyToClassName={}},e.prototype._getStyleElement=function(){var t=this;return!this._styleElement&&typeof document<"u"&&(this._styleElement=this._createStyleElement(),bG||window.requestAnimationFrame(function(){t._styleElement=void 0})),this._styleElement},e.prototype._createStyleElement=function(){var t=document.head,n=document.createElement("style"),r=null;n.setAttribute("data-merge-styles","true");var o=this._config.cspSettings;if(o&&o.nonce&&n.setAttribute("nonce",o.nonce),this._lastStyleElement)r=this._lastStyleElement.nextElementSibling;else{var i=this._findPlaceholderStyleTag();i?r=i.nextElementSibling:r=t.childNodes[0]}return t.insertBefore(n,t.contains(r)?r:null),this._lastStyleElement=n,n},e.prototype._findPlaceholderStyleTag=function(){var t=document.head;return t?t.querySelector("style[data-merge-styles]"):null},e}();function mF(){for(var e=[],t=0;t<arguments.length;t++)e[t]=arguments[t];var n=[],r=[],o=Qi.getInstance();function i(a){for(var s=0,l=a;s<l.length;s++){var u=l[s];if(u)if(typeof u=="string")if(u.indexOf(" ")>=0)i(u.split(" "));else{var d=o.argsFromClassName(u);d?i(d):n.indexOf(u)===-1&&n.push(u)}else Array.isArray(u)?i(u):typeof u=="object"&&r.push(u)}}return i(e),{classes:n,objects:r}}function gF(e){cd!==e&&(cd=e)}function vF(){return cd===void 0&&(cd=typeof document<"u"&&!!document.documentElement&&document.documentElement.getAttribute("dir")==="rtl"),cd}var cd;cd=vF();function Eb(){return{rtl:vF()}}var Y8={};function yG(e,t){var n=e[t];n.charAt(0)!=="-"&&(e[t]=Y8[n]=Y8[n]||n.replace(/([A-Z])/g,"-$1").toLowerCase())}var vm;function EG(){var e;if(!vm){var t=typeof document<"u"?document:void 0,n=typeof navigator<"u"?navigator:void 0,r=(e=n==null?void 0:n.userAgent)===null||e===void 0?void 0:e.toLowerCase();t?vm={isWebkit:!!(t&&"WebkitAppearance"in t.documentElement.style),isMoz:!!(r&&r.indexOf("firefox")>-1),isOpera:!!(r&&r.indexOf("opera")>-1),isMs:!!(n&&(/rv:11.0/i.test(n.userAgent)||/Edge\/\d./i.test(navigator.userAgent)))}:vm={isWebkit:!0,isMoz:!0,isOpera:!0,isMs:!0}}return vm}var Q8={"user-select":1};function _G(e,t){var n=EG(),r=e[t];if(Q8[r]){var o=e[t+1];Q8[r]&&(n.isWebkit&&e.push("-webkit-"+r,o),n.isMoz&&e.push("-moz-"+r,o),n.isMs&&e.push("-ms-"+r,o),n.isOpera&&e.push("-o-"+r,o))}}var TG=["column-count","font-weight","flex","flex-grow","flex-shrink","fill-opacity","opacity","order","z-index","zoom"];function wG(e,t){var n=e[t],r=e[t+1];if(typeof r=="number"){var o=TG.indexOf(n)>-1,i=n.indexOf("--")>-1,a=o||i?"":"px";e[t+1]=""+r+a}}var bm,Wl="left",Gl="right",kG="@noflip",X8=(bm={},bm[Wl]=Gl,bm[Gl]=Wl,bm),Z8={"w-resize":"e-resize","sw-resize":"se-resize","nw-resize":"ne-resize"};function SG(e,t,n){if(e.rtl){var r=t[n];if(!r)return;var o=t[n+1];if(typeof o=="string"&&o.indexOf(kG)>=0)t[n+1]=o.replace(/\s*(?:\/\*\s*)?\@noflip\b(?:\s*\*\/)?\s*?/g,"");else if(r.indexOf(Wl)>=0)t[n]=r.replace(Wl,Gl);else if(r.indexOf(Gl)>=0)t[n]=r.replace(Gl,Wl);else if(String(o).indexOf(Wl)>=0)t[n+1]=o.replace(Wl,Gl);else if(String(o).indexOf(Gl)>=0)t[n+1]=o.replace(Gl,Wl);else if(X8[r])t[n]=X8[r];else if(Z8[o])t[n+1]=Z8[o];else switch(r){case"margin":case"padding":t[n+1]=CG(o);break;case"box-shadow":t[n+1]=xG(o,0);break}}}function xG(e,t){var n=e.split(" "),r=parseInt(n[t],10);return n[0]=n[0].replace(String(r),String(r*-1)),n.join(" ")}function CG(e){if(typeof e=="string"){var t=e.split(" ");if(t.length===4)return t[0]+" "+t[3]+" "+t[2]+" "+t[1]}return e}function AG(e){for(var t=[],n=0,r=0,o=0;o<e.length;o++)switch(e[o]){case"(":r++;break;case")":r&&r--;break;case"	":case" ":r||(o>n&&t.push(e.substring(n,o)),n=o+1);break}return n<e.length&&t.push(e.substring(n)),t}var NG="displayName";function FG(e){var t=e&&e["&"];return t?t.displayName:void 0}var bF=/\:global\((.+?)\)/g;function IG(e){if(!bF.test(e))return e;for(var t=[],n=/\:global\((.+?)\)/g,r=null;r=n.exec(e);)r[1].indexOf(",")>-1&&t.push([r.index,r.index+r[0].length,r[1].split(",").map(function(o){return":global("+o.trim()+")"}).join(", ")]);return t.reverse().reduce(function(o,i){var a=i[0],s=i[1],l=i[2],u=o.slice(0,a),d=o.slice(s);return u+l+d},e)}function J8(e,t){return e.indexOf(":global(")>=0?e.replace(bF,"$1"):e.indexOf(":")===0?t+e:e.indexOf("&")<0?t+" "+e:e}function ex(e,t,n,r){t===void 0&&(t={__order:[]}),n.indexOf("@")===0?(n=n+"{"+e,fd([r],t,n)):n.indexOf(",")>-1?IG(n).split(",").map(function(o){return o.trim()}).forEach(function(o){return fd([r],t,J8(o,e))}):fd([r],t,J8(n,e))}function fd(e,t,n){t===void 0&&(t={__order:[]}),n===void 0&&(n="&");var r=Qi.getInstance(),o=t[n];o||(o={},t[n]=o,t.__order.push(n));for(var i=0,a=e;i<a.length;i++){var s=a[i];if(typeof s=="string"){var l=r.argsFromClassName(s);l&&fd(l,t,n)}else if(Array.isArray(s))fd(s,t,n);else for(var u in s)if(s.hasOwnProperty(u)){var d=s[u];if(u==="selectors"){var h=s.selectors;for(var p in h)h.hasOwnProperty(p)&&ex(n,t,p,h[p])}else typeof d=="object"?d!==null&&ex(n,t,u,d):d!==void 0&&(u==="margin"||u==="padding"?BG(o,u,d):o[u]=d)}}return t}function BG(e,t,n){var r=typeof n=="string"?AG(n):[n];r.length===0&&r.push(n),r[r.length-1]==="!important"&&(r=r.slice(0,-1).map(function(o){return o+" !important"})),e[t+"Top"]=r[0],e[t+"Right"]=r[1]||r[0],e[t+"Bottom"]=r[2]||r[0],e[t+"Left"]=r[3]||r[1]||r[0]}function RG(e,t){for(var n=[e.rtl?"rtl":"ltr"],r=!1,o=0,i=t.__order;o<i.length;o++){var a=i[o];n.push(a);var s=t[a];for(var l in s)s.hasOwnProperty(l)&&s[l]!==void 0&&(r=!0,n.push(l,s[l]))}return r?n.join(""):void 0}function yF(e,t){return t<=0?"":t===1?e:e+yF(e,t-1)}function _4(e,t){if(!t)return"";var n=[];for(var r in t)t.hasOwnProperty(r)&&r!==NG&&t[r]!==void 0&&n.push(r,t[r]);for(var o=0;o<n.length;o+=2)yG(n,o),wG(n,o),SG(e,n,o),_G(n,o);for(var o=1;o<n.length;o+=4)n.splice(o,1,":",n[o],";");return n.join("")}function EF(e){for(var t=[],n=1;n<arguments.length;n++)t[n-1]=arguments[n];var r=fd(t),o=RG(e,r);if(o){var i=Qi.getInstance(),a={className:i.classNameFromKey(o),key:o,args:t};if(!a.className){a.className=i.getClassName(FG(r));for(var s=[],l=0,u=r.__order;l<u.length;l++){var d=u[l];s.push(d,_4(e,r[d]))}a.rulesToInsert=s}return a}}function _F(e,t){t===void 0&&(t=1);var n=Qi.getInstance(),r=e.className,o=e.key,i=e.args,a=e.rulesToInsert;if(a){for(var s=0;s<a.length;s+=2){var l=a[s+1];if(l){var u=a[s];u=u.replace(/&/g,yF("."+e.className,t));var d=u+"{"+l+"}"+(u.indexOf("@")===0?"}":"");n.insertRule(d)}}n.cacheClassName(r,o,i,a)}}function OG(e){for(var t=[],n=1;n<arguments.length;n++)t[n-1]=arguments[n];var r=EF.apply(void 0,Yi([e],t));return r?(_F(r,e.specificityMultiplier),r.className):""}function Fo(){for(var e=[],t=0;t<arguments.length;t++)e[t]=arguments[t];return TF(e,Eb())}function TF(e,t){var n=e instanceof Array?e:[e],r=mF(n),o=r.classes,i=r.objects;return i.length&&o.push(OG(t||{},i)),o.join(" ")}function jc(){for(var e=[],t=0;t<arguments.length;t++)e[t]=arguments[t];if(e&&e.length===1&&e[0]&&!e[0].subComponentStyles)return e[0];for(var n={},r={},o=0,i=e;o<i.length;o++){var a=i[o];if(a){for(var s in a)if(a.hasOwnProperty(s)){if(s==="subComponentStyles"&&a.subComponentStyles!==void 0){var l=a.subComponentStyles;for(var u in l)l.hasOwnProperty(u)&&(r.hasOwnProperty(u)?r[u].push(l[u]):r[u]=[l[u]]);continue}var d=n[s],h=a[s];d===void 0?n[s]=h:n[s]=Yi(Yi([],Array.isArray(d)?d:[d]),Array.isArray(h)?h:[h])}}}if(Object.keys(r).length>0){n.subComponentStyles={};var p=n.subComponentStyles,m=function(v){if(r.hasOwnProperty(v)){var _=r[v];p[v]=function(b){return jc.apply(void 0,_.map(function(E){return typeof E=="function"?E(b):E}))}}};for(var u in r)m(u)}return n}function q0(){for(var e=[],t=0;t<arguments.length;t++)e[t]=arguments[t];return wF(e,Eb())}function wF(e,t){var n={subComponentStyles:{}},r=e[0];if(!r&&e.length<=1)return{subComponentStyles:{}};var o=jc.apply(void 0,e),i=[];for(var a in o)if(o.hasOwnProperty(a)){if(a==="subComponentStyles"){n.subComponentStyles=o.subComponentStyles||{};continue}var s=o[a],l=mF(s),u=l.classes,d=l.objects;if(d!=null&&d.length){var h=EF(t||{},{displayName:a},d);h&&(i.push(h),n[a]=u.concat([h.className]).join(" "))}else n[a]=u.join(" ")}for(var p=0,m=i;p<m.length;p++){var h=m[p];h&&_F(h,t==null?void 0:t.specificityMultiplier)}return n}function kF(e){for(var t=[],n=1;n<arguments.length;n++)t[n-1]=arguments[n];for(var r=[],o=0,i=t;o<i.length;o++){var a=i[o];a&&r.push(typeof a=="function"?a(e):a)}return r.length===1?r[0]:r.length?jc.apply(void 0,r):{}}function SF(e){var t=Qi.getInstance(),n=_4(Eb(),e),r=t.classNameFromKey(n);if(!r){var o=t.getClassName();t.insertRule("@font-face{"+n+"}",!0),t.cacheClassName(o,n,[],["font-face",n])}}function Ji(e){var t=Qi.getInstance(),n=[];for(var r in e)e.hasOwnProperty(r)&&n.push(r,"{",_4(Eb(),e[r]),"}");var o=n.join(""),i=t.classNameFromKey(o);if(i)return i;var a=t.getClassName();return t.insertRule("@keyframes "+a+"{"+o+"}",!0),t.cacheClassName(a,o,[],["keyframes",o]),a}function DG(e){var t={},n=function(o){if(e.hasOwnProperty(o)){var i;Object.defineProperty(t,o,{get:function(){return i===void 0&&(i=Fo(e[o]).toString()),i},enumerable:!0,configurable:!0})}};for(var r in e)n(r);return t}var gE=void 0;try{gE=window}catch{}function ur(e){if(!(typeof gE>"u")){var t=e;return t&&t.ownerDocument&&t.ownerDocument.defaultView?t.ownerDocument.defaultView:gE}}var _b=function(){function e(t,n){this._timeoutIds=null,this._immediateIds=null,this._intervalIds=null,this._animationFrameIds=null,this._isDisposed=!1,this._parent=t||null,this._onErrorHandler=n,this._noop=function(){}}return e.prototype.dispose=function(){var t;if(this._isDisposed=!0,this._parent=null,this._timeoutIds){for(t in this._timeoutIds)this._timeoutIds.hasOwnProperty(t)&&this.clearTimeout(parseInt(t,10));this._timeoutIds=null}if(this._immediateIds){for(t in this._immediateIds)this._immediateIds.hasOwnProperty(t)&&this.clearImmediate(parseInt(t,10));this._immediateIds=null}if(this._intervalIds){for(t in this._intervalIds)this._intervalIds.hasOwnProperty(t)&&this.clearInterval(parseInt(t,10));this._intervalIds=null}if(this._animationFrameIds){for(t in this._animationFrameIds)this._animationFrameIds.hasOwnProperty(t)&&this.cancelAnimationFrame(parseInt(t,10));this._animationFrameIds=null}},e.prototype.setTimeout=function(t,n){var r=this,o=0;return this._isDisposed||(this._timeoutIds||(this._timeoutIds={}),o=setTimeout(function(){try{r._timeoutIds&&delete r._timeoutIds[o],t.apply(r._parent)}catch(i){r._logError(i)}},n),this._timeoutIds[o]=!0),o},e.prototype.clearTimeout=function(t){this._timeoutIds&&this._timeoutIds[t]&&(clearTimeout(t),delete this._timeoutIds[t])},e.prototype.setImmediate=function(t,n){var r=this,o=0,i=ur(n);if(!this._isDisposed){this._immediateIds||(this._immediateIds={});var a=function(){try{r._immediateIds&&delete r._immediateIds[o],t.apply(r._parent)}catch(s){r._logError(s)}};o=i.setTimeout(a,0),this._immediateIds[o]=!0}return o},e.prototype.clearImmediate=function(t,n){var r=ur(n);this._immediateIds&&this._immediateIds[t]&&(r.clearTimeout(t),delete this._immediateIds[t])},e.prototype.setInterval=function(t,n){var r=this,o=0;return this._isDisposed||(this._intervalIds||(this._intervalIds={}),o=setInterval(function(){try{t.apply(r._parent)}catch(i){r._logError(i)}},n),this._intervalIds[o]=!0),o},e.prototype.clearInterval=function(t){this._intervalIds&&this._intervalIds[t]&&(clearInterval(t),delete this._intervalIds[t])},e.prototype.throttle=function(t,n,r){var o=this;if(this._isDisposed)return this._noop;var i=n||0,a=!0,s=!0,l=0,u,d,h=null;r&&typeof r.leading=="boolean"&&(a=r.leading),r&&typeof r.trailing=="boolean"&&(s=r.trailing);var p=function(v){var _=Date.now(),b=_-l,E=a?i-b:i;return b>=i&&(!v||a)?(l=_,h&&(o.clearTimeout(h),h=null),u=t.apply(o._parent,d)):h===null&&s&&(h=o.setTimeout(p,E)),u},m=function(){for(var v=[],_=0;_<arguments.length;_++)v[_]=arguments[_];return d=v,p(!0)};return m},e.prototype.debounce=function(t,n,r){var o=this;if(this._isDisposed){var i=function(){};return i.cancel=function(){},i.flush=function(){return null},i.pending=function(){return!1},i}var a=n||0,s=!1,l=!0,u=null,d=0,h=Date.now(),p,m,v=null;r&&typeof r.leading=="boolean"&&(s=r.leading),r&&typeof r.trailing=="boolean"&&(l=r.trailing),r&&typeof r.maxWait=="number"&&!isNaN(r.maxWait)&&(u=r.maxWait);var _=function(C){v&&(o.clearTimeout(v),v=null),h=C},b=function(C){_(C),p=t.apply(o._parent,m)},E=function(C){var A=Date.now(),P=!1;C&&(s&&A-d>=a&&(P=!0),d=A);var I=A-d,j=a-I,H=A-h,K=!1;return u!==null&&(H>=u&&v?K=!0:j=Math.min(j,u-H)),I>=a||K||P?b(A):(v===null||!C)&&l&&(v=o.setTimeout(E,j)),p},w=function(){return!!v},k=function(){w()&&_(Date.now())},y=function(){return w()&&b(Date.now()),p},F=function(){for(var C=[],A=0;A<arguments.length;A++)C[A]=arguments[A];return m=C,E(!0)};return F.cancel=k,F.flush=y,F.pending=w,F},e.prototype.requestAnimationFrame=function(t,n){var r=this,o=0,i=ur(n);if(!this._isDisposed){this._animationFrameIds||(this._animationFrameIds={});var a=function(){try{r._animationFrameIds&&delete r._animationFrameIds[o],t.apply(r._parent)}catch(s){r._logError(s)}};o=i.requestAnimationFrame?i.requestAnimationFrame(a):i.setTimeout(a,0),this._animationFrameIds[o]=!0}return o},e.prototype.cancelAnimationFrame=function(t,n){var r=ur(n);this._animationFrameIds&&this._animationFrameIds[t]&&(r.cancelAnimationFrame?r.cancelAnimationFrame(t):r.clearTimeout(t),delete this._animationFrameIds[t])},e.prototype._logError=function(t){this._onErrorHandler&&this._onErrorHandler(t)},e}();function T4(e,t){for(var n in e)if(e.hasOwnProperty(n)&&(!t.hasOwnProperty(n)||t[n]!==e[n]))return!1;for(var n in t)if(t.hasOwnProperty(n)&&!e.hasOwnProperty(n))return!1;return!0}function pc(e){for(var t=[],n=1;n<arguments.length;n++)t[n-1]=arguments[n];return PG.apply(this,[null,e].concat(t))}function PG(e,t){for(var n=[],r=2;r<arguments.length;r++)n[r-2]=arguments[r];t=t||{};for(var o=0,i=n;o<i.length;o++){var a=i[o];if(a)for(var s in a)a.hasOwnProperty(s)&&(!e||e(s))&&(t[s]=a[s])}return t}var Ws=function(){function e(t){this._id=e._uniqueId++,this._parent=t,this._eventRecords=[]}return e.raise=function(t,n,r,o){var i;if(e._isElement(t)){if(typeof document<"u"&&document.createEvent){var a=document.createEvent("HTMLEvents");a.initEvent(n,o||!1,!0),pc(a,r),i=t.dispatchEvent(a)}else if(typeof document<"u"&&document.createEventObject){var s=document.createEventObject(r);t.fireEvent("on"+n,s)}}else for(;t&&i!==!1;){var l=t.__events__,u=l?l[n]:null;if(u){for(var d in u)if(u.hasOwnProperty(d))for(var h=u[d],p=0;i!==!1&&p<h.length;p++){var m=h[p];m.objectCallback&&(i=m.objectCallback.call(m.parent,r))}}t=o?t.parent:null}return i},e.isObserved=function(t,n){var r=t&&t.__events__;return!!r&&!!r[n]},e.isDeclared=function(t,n){var r=t&&t.__declaredEvents;return!!r&&!!r[n]},e.stopPropagation=function(t){t.stopPropagation?t.stopPropagation():t.cancelBubble=!0},e._isElement=function(t){return!!t&&(!!t.addEventListener||typeof HTMLElement<"u"&&t instanceof HTMLElement)},e.prototype.dispose=function(){this._isDisposed||(this._isDisposed=!0,this.off(),this._parent=null)},e.prototype.onAll=function(t,n,r){for(var o in n)n.hasOwnProperty(o)&&this.on(t,o,n[o],r)},e.prototype.on=function(t,n,r,o){var i=this;if(n.indexOf(",")>-1)for(var a=n.split(/[ ,]+/),s=0;s<a.length;s++)this.on(t,a[s],r,o);else{var l=this._parent,u={target:t,eventName:n,parent:l,callback:r,options:o},a=t.__events__=t.__events__||{};if(a[n]=a[n]||{count:0},a[n][this._id]=a[n][this._id]||[],a[n][this._id].push(u),a[n].count++,e._isElement(t)){var d=function(){for(var m=[],v=0;v<arguments.length;v++)m[v]=arguments[v];if(!i._isDisposed){var _;try{if(_=r.apply(l,m),_===!1&&m[0]){var b=m[0];b.preventDefault&&b.preventDefault(),b.stopPropagation&&b.stopPropagation(),b.cancelBubble=!0}}catch{}return _}};u.elementCallback=d,t.addEventListener?t.addEventListener(n,d,o):t.attachEvent&&t.attachEvent("on"+n,d)}else{var h=function(){for(var m=[],v=0;v<arguments.length;v++)m[v]=arguments[v];if(!i._isDisposed)return r.apply(l,m)};u.objectCallback=h}this._eventRecords.push(u)}},e.prototype.off=function(t,n,r,o){for(var i=0;i<this._eventRecords.length;i++){var a=this._eventRecords[i];if((!t||t===a.target)&&(!n||n===a.eventName)&&(!r||r===a.callback)&&(typeof o!="boolean"||o===a.options)){var s=a.target.__events__,l=s[a.eventName],u=l?l[this._id]:null;u&&(u.length===1||!r?(l.count-=u.length,delete s[a.eventName][this._id]):(l.count--,u.splice(u.indexOf(a),1)),l.count||delete s[a.eventName]),a.elementCallback&&(a.target.removeEventListener?a.target.removeEventListener(a.eventName,a.elementCallback,a.options):a.target.detachEvent&&a.target.detachEvent("on"+a.eventName,a.elementCallback)),this._eventRecords.splice(i--,1)}}},e.prototype.raise=function(t,n,r){return e.raise(this._parent,t,n,r)},e.prototype.declare=function(t){var n=this._parent.__declaredEvents=this._parent.__declaredEvents||{};if(typeof t=="string")n[t]=!0;else for(var r=0;r<t.length;r++)n[t[r]]=!0},e._uniqueId=0,e}();function ss(e){if(!(typeof document>"u")){var t=e;return t&&t.ownerDocument?t.ownerDocument:document}}var Z5;Fo({overflow:"hidden !important"});function MG(){if(Z5===void 0){var e=document.createElement("div");e.style.setProperty("width","100px"),e.style.setProperty("height","100px"),e.style.setProperty("overflow","scroll"),e.style.setProperty("position","absolute"),e.style.setProperty("top","-9999px"),document.body.appendChild(e),Z5=e.offsetWidth-e.clientWidth,document.body.removeChild(e)}return Z5}var LG=void 0;function xF(e){console&&console.warn&&console.warn(e)}function CF(e,t,n,r,o){if(o===!0&&!1)for(var i,a;i<a.length;i++)var s}(function(e){yi(t,e);function t(n,r){var o=e.call(this,n,r)||this;return jG(o,t.prototype,["componentDidMount","shouldComponentUpdate","getSnapshotBeforeUpdate","render","componentDidUpdate","componentWillUnmount"]),o}return t.prototype.componentDidUpdate=function(n,r){this._updateComponentRef(n,this.props)},t.prototype.componentDidMount=function(){this._setComponentRef(this.props.componentRef,this)},t.prototype.componentWillUnmount=function(){if(this._setComponentRef(this.props.componentRef,null),this.__disposables){for(var n=0,r=this._disposables.length;n<r;n++){var o=this.__disposables[n];o.dispose&&o.dispose()}this.__disposables=null}},Object.defineProperty(t.prototype,"className",{get:function(){if(!this.__className){var n=/function (.{1,})\(/,r=n.exec(this.constructor.toString());this.__className=r&&r.length>1?r[1]:""}return this.__className},enumerable:!1,configurable:!0}),Object.defineProperty(t.prototype,"_disposables",{get:function(){return this.__disposables||(this.__disposables=[]),this.__disposables},enumerable:!1,configurable:!0}),Object.defineProperty(t.prototype,"_async",{get:function(){return this.__async||(this.__async=new _b(this),this._disposables.push(this.__async)),this.__async},enumerable:!1,configurable:!0}),Object.defineProperty(t.prototype,"_events",{get:function(){return this.__events||(this.__events=new Ws(this),this._disposables.push(this.__events)),this.__events},enumerable:!1,configurable:!0}),t.prototype._resolveRef=function(n){var r=this;return this.__resolves||(this.__resolves={}),this.__resolves[n]||(this.__resolves[n]=function(o){return r[n]=o}),this.__resolves[n]},t.prototype._updateComponentRef=function(n,r){r===void 0&&(r={}),n&&r&&n.componentRef!==r.componentRef&&(this._setComponentRef(n.componentRef,null),this._setComponentRef(r.componentRef,this))},t.prototype._warnDeprecations=function(n){this.className,this.props},t.prototype._warnMutuallyExclusive=function(n){this.className,this.props},t.prototype._warnConditionallyRequiredProps=function(n,r,o){CF(this.className,this.props,n,r,o)},t.prototype._setComponentRef=function(n,r){!this._skipComponentRefResolution&&n&&(typeof n=="function"&&n(r),typeof n=="object"&&(n.current=r))},t})(T.Component);function jG(e,t,n){for(var r=0,o=n.length;r<o;r++)zG(e,t,n[r])}function zG(e,t,n){var r=e[n],o=t[n];(r||o)&&(e[n]=function(){for(var i=[],a=0;a<arguments.length;a++)i[a]=arguments[a];var s;return o&&(s=o.apply(this,i)),r!==o&&(s=r.apply(this,i)),s})}function AF(){return null}var J5="__globalSettings__",w4="__callbacks__",HG=0,NF=function(){function e(){}return e.getValue=function(t,n){var r=vE();return r[t]===void 0&&(r[t]=typeof n=="function"?n():n),r[t]},e.setValue=function(t,n){var r=vE(),o=r[w4],i=r[t];if(n!==i){r[t]=n;var a={oldValue:i,value:n,key:t};for(var s in o)o.hasOwnProperty(s)&&o[s](a)}return n},e.addChangeListener=function(t){var n=t.__id__,r=tx();n||(n=t.__id__=String(HG++)),r[n]=t},e.removeChangeListener=function(t){var n=tx();delete n[t.__id__]},e}();function vE(){var e,t=ur(),n=t||{};return n[J5]||(n[J5]=(e={},e[w4]={},e)),n[J5]}function tx(){var e=vE();return e[w4]}var qt={backspace:8,tab:9,enter:13,shift:16,ctrl:17,alt:18,pauseBreak:19,capslock:20,escape:27,space:32,pageUp:33,pageDown:34,end:35,home:36,left:37,up:38,right:39,down:40,insert:45,del:46,zero:48,one:49,two:50,three:51,four:52,five:53,six:54,seven:55,eight:56,nine:57,colon:58,a:65,b:66,c:67,d:68,e:69,f:70,g:71,h:72,i:73,j:74,k:75,l:76,m:77,n:78,o:79,p:80,q:81,r:82,s:83,t:84,u:85,v:86,w:87,x:88,y:89,z:90,leftWindow:91,rightWindow:92,select:93,zero_numpad:96,one_numpad:97,two_numpad:98,three_numpad:99,four_numpad:100,five_numpad:101,six_numpad:102,seven_numpad:103,eight_numpad:104,nine_numpad:105,multiply:106,add:107,subtract:109,decimalPoint:110,divide:111,f1:112,f2:113,f3:114,f4:115,f5:116,f6:117,f7:118,f8:119,f9:120,f10:121,f11:122,f12:123,numlock:144,scrollLock:145,semicolon:186,equalSign:187,comma:188,dash:189,period:190,forwardSlash:191,graveAccent:192,openBracket:219,backSlash:220,closeBracket:221,singleQuote:222},ts=function(){function e(t,n,r,o){t===void 0&&(t=0),n===void 0&&(n=0),r===void 0&&(r=0),o===void 0&&(o=0),this.top=r,this.bottom=o,this.left=t,this.right=n}return Object.defineProperty(e.prototype,"width",{get:function(){return this.right-this.left},enumerable:!1,configurable:!0}),Object.defineProperty(e.prototype,"height",{get:function(){return this.bottom-this.top},enumerable:!1,configurable:!0}),e.prototype.equals=function(t){return parseFloat(this.top.toFixed(4))===parseFloat(t.top.toFixed(4))&&parseFloat(this.bottom.toFixed(4))===parseFloat(t.bottom.toFixed(4))&&parseFloat(this.left.toFixed(4))===parseFloat(t.left.toFixed(4))&&parseFloat(this.right.toFixed(4))===parseFloat(t.right.toFixed(4))},e}();function UG(e){for(var t=[],n=1;n<arguments.length;n++)t[n-1]=arguments[n];return t.length<2?t[0]:function(){for(var r=[],o=0;o<arguments.length;o++)r[o]=arguments[o];t.forEach(function(i){return i&&i.apply(e,r)})}}function $0(){for(var e=[],t=0;t<arguments.length;t++)e[t]=arguments[t];var n=e.filter(function(r){return r}).join(" ").trim();return n===""?void 0:n}function qG(e,t,n){var r=e.slice();return r.splice(t,0,n),r}function $G(e,t){if(e.length!==t.length)return!1;for(var n=0;n<e.length;n++)if(e[n]!==t[n])return!1;return!0}function FF(e){var t=null;try{var n=ur();t=n?n.sessionStorage.getItem(e):null}catch{}return t}function WG(e,t){var n;try{(n=ur())===null||n===void 0||n.sessionStorage.setItem(e,t)}catch{}}var IF="isRTL",Hs;function ls(e){if(e===void 0&&(e={}),e.rtl!==void 0)return e.rtl;if(Hs===void 0){var t=FF(IF);t!==null&&(Hs=t==="1",GG(Hs));var n=ss();Hs===void 0&&n&&(Hs=(n.body&&n.body.getAttribute("dir")||n.documentElement.getAttribute("dir"))==="rtl",gF(Hs))}return!!Hs}function GG(e,t){t===void 0&&(t=!1);var n=ss();n&&n.documentElement.setAttribute("dir",e?"rtl":"ltr"),t&&WG(IF,e?"1":"0"),Hs=e,gF(Hs)}function KG(e){return e&&!!e._virtual}function VG(e){var t;return e&&KG(e)&&(t=e._virtual.parent),t}function BF(e,t){return t===void 0&&(t=!0),e&&(t&&VG(e)||e.parentNode&&e.parentNode)}function bE(e,t,n){n===void 0&&(n=!0);var r=!1;if(e&&t)if(n)if(e===t)r=!0;else for(r=!1;t;){var o=BF(t);if(o===e){r=!0;break}t=o}else e.contains&&(r=e.contains(t));return r}function k4(e,t){return!e||e===document.body?null:t(e)?e:k4(BF(e),t)}function YG(e,t){var n=k4(e,function(r){return r.hasAttribute(t)});return n&&n.getAttribute(t)}var yE="data-portal-element";function QG(e){e.setAttribute(yE,"true")}function XG(e,t){var n=k4(e,function(r){return t===r||r.hasAttribute(yE)});return n!==null&&n.hasAttribute(yE)}function ZG(e,t){var n=e,r=t;n._virtual||(n._virtual={children:[]});var o=n._virtual.parent;if(o&&o!==t){var i=o._virtual.children.indexOf(n);i>-1&&o._virtual.children.splice(i,1)}n._virtual.parent=r||void 0,r&&(r._virtual||(r._virtual={children:[]}),r._virtual.children.push(n))}var JG="data-is-focusable",eK="data-is-visible",tK="data-focuszone-id",nK="data-is-sub-focuszone";function rK(e,t,n){return xh(e,t,!0,!1,!1,n)}function oK(e,t,n){return jf(e,t,!0,!1,!0,n)}function iK(e){var t=xh(e,e,!0,!1,!1,!0);return t?(lK(t),!0):!1}function jf(e,t,n,r,o,i,a,s){if(!t||!a&&t===e)return null;var l=RF(t);if(o&&l&&(i||!(OF(t)||DF(t)))){var u=jf(e,t.lastElementChild,!0,!0,!0,i,a,s);if(u){if(s&&EE(u,!0)||!s)return u;var d=jf(e,u.previousElementSibling,!0,!0,!0,i,a,s);if(d)return d;for(var h=u.parentElement;h&&h!==t;){var p=jf(e,h.previousElementSibling,!0,!0,!0,i,a,s);if(p)return p;h=h.parentElement}}}if(n&&l&&EE(t,s))return t;var m=jf(e,t.previousElementSibling,!0,!0,!0,i,a,s);return m||(r?null:jf(e,t.parentElement,!0,!1,!1,i,a,s))}function xh(e,t,n,r,o,i,a,s){if(!t||t===e&&o&&!a)return null;var l=RF(t);if(n&&l&&EE(t,s))return t;if(!o&&l&&(i||!(OF(t)||DF(t)))){var u=xh(e,t.firstElementChild,!0,!0,!1,i,a,s);if(u)return u}if(t===e)return null;var d=xh(e,t.nextElementSibling,!0,!0,!1,i,a,s);return d||(r?null:xh(e,t.parentElement,!1,!1,!0,i,a,s))}function RF(e){if(!e||!e.getAttribute)return!1;var t=e.getAttribute(eK);return t!=null?t==="true":e.offsetHeight!==0||e.offsetParent!==null||e.isVisible===!0}function EE(e,t){if(!e||e.disabled)return!1;var n=0,r=null;e&&e.getAttribute&&(r=e.getAttribute("tabIndex"),r&&(n=parseInt(r,10)));var o=e.getAttribute?e.getAttribute(JG):null,i=r!==null&&n>=0,a=!!e&&o!=="false"&&(e.tagName==="A"||e.tagName==="BUTTON"||e.tagName==="INPUT"||e.tagName==="TEXTAREA"||e.tagName==="SELECT"||o==="true"||i);return t?n!==-1&&a:a}function OF(e){return!!(e&&e.getAttribute&&e.getAttribute(tK))}function DF(e){return!!(e&&e.getAttribute&&e.getAttribute(nK)==="true")}function aK(e){var t=ss(e),n=t&&t.activeElement;return!!(n&&bE(e,n))}function sK(e,t){return YG(e,t)!=="true"}var bf=void 0;function lK(e){if(e){if(bf){bf=e;return}bf=e;var t=ur(e);t&&t.requestAnimationFrame(function(){bf&&bf.focus(),bf=void 0})}}function zf(e,t,n,r){return e.addEventListener(t,n,r),function(){return e.removeEventListener(t,n,r)}}var uK=50,cK=5,yg=0,e2=Qi.getInstance();e2&&e2.onReset&&e2.onReset(function(){return yg++});var ym="__retval__";function ml(e){e===void 0&&(e={});var t=new Map,n=0,r=0,o=yg,i=function(a,s){var l;if(s===void 0&&(s={}),e.useStaticStyles&&typeof a=="function"&&a.__noStyleOverride__)return a(s);r++;var u=t,d=s.theme,h=d&&d.rtl!==void 0?d.rtl:ls(),p=e.disableCaching;if(o!==yg&&(o=yg,t=new Map,n=0),e.disableCaching||(u=nx(t,a),u=nx(u,s)),(p||!u[ym])&&(a===void 0?u[ym]={}:u[ym]=wF([typeof a=="function"?a(s):a],{rtl:!!h,specificityMultiplier:e.useStaticStyles?cK:void 0}),p||n++),n>(e.cacheSize||uK)){var m=ur();!((l=m==null?void 0:m.FabricConfig)===null||l===void 0)&&l.enableClassNameCacheFullWarning&&(console.warn("Styles are being recalculated too frequently. Cache miss rate is "+n+"/"+r+"."),console.trace()),t.clear(),n=0,e.disableCaching=!0}return u[ym]};return i}function t2(e,t){return t=fK(t),e.has(t)||e.set(t,new Map),e.get(t)}function nx(e,t){if(typeof t=="function"){var n=t.__cachedInputs__;if(n)for(var r=0,o=t.__cachedInputs__;r<o.length;r++){var i=o[r];e=t2(e,i)}else e=t2(e,t)}else if(typeof t=="object")for(var a in t)t.hasOwnProperty(a)&&(e=t2(e,t[a]));return e}function fK(e){switch(e){case void 0:return"__undefined__";case null:return"__null__";default:return e}}var rx=!1,Eg=0,dK={empty:!0},n2={},r0=typeof WeakMap>"u"?null:WeakMap;function hK(){Eg++}function Cr(e,t,n){if(t===void 0&&(t=100),n===void 0&&(n=!1),!r0)return e;if(!rx){var r=Qi.getInstance();r&&r.onReset&&Qi.getInstance().onReset(hK),rx=!0}var o,i=0,a=Eg;return function(){for(var l=[],u=0;u<arguments.length;u++)l[u]=arguments[u];var d=o;(o===void 0||a!==Eg||t>0&&i>t)&&(o=ox(),i=0,a=Eg),d=o;for(var h=0;h<l.length;h++){var p=pK(l[h]);d.map.has(p)||d.map.set(p,ox()),d=d.map.get(p)}return d.hasOwnProperty("value")||(d.value=e.apply(void 0,l),i++),n&&(d.value===null||d.value===void 0)&&(d.value=e.apply(void 0,l)),d.value}}function Od(e){if(!r0)return e;var t=new r0;function n(r){if(!r||typeof r!="function"&&typeof r!="object")return e(r);if(t.has(r))return t.get(r);var o=e(r);return t.set(r,o),o}return n}function pK(e){if(e){if(typeof e=="object"||typeof e=="function")return e;n2[e]||(n2[e]={val:e})}else return dK;return n2[e]}function ox(){return{map:r0?new r0:null}}function mK(e){var t=e,n=Od(function(r){if(e===r)throw new Error("Attempted to compose a component with itself.");var o=r,i=Od(function(s){var l=function(u){return T.createElement(o,oe({},u,{defaultRender:s}))};return l}),a=function(s){var l=s.defaultRender;return T.createElement(t,oe({},s,{defaultRender:l?i(l):o}))};return a});return n}var gK=Od(mK);function PF(e,t){return gK(e)(t)}function Ys(){for(var e=[],t=0;t<arguments.length;t++)e[t]=arguments[t];for(var n=[],r=0,o=e;r<o.length;r++){var i=o[r];if(i)if(typeof i=="string")n.push(i);else if(i.hasOwnProperty("toString")&&typeof i.toString=="function")n.push(i.toString());else for(var a in i)i[a]&&n.push(a)}return n.join(" ")}var vK="customizations",bK={settings:{},scopedSettings:{},inCustomizerContext:!1},Ol=NF.getValue(vK,{settings:{},scopedSettings:{},inCustomizerContext:!1}),Em=[],$i=function(){function e(){}return e.reset=function(){Ol.settings={},Ol.scopedSettings={}},e.applySettings=function(t){Ol.settings=oe(oe({},Ol.settings),t),e._raiseChange()},e.applyScopedSettings=function(t,n){Ol.scopedSettings[t]=oe(oe({},Ol.scopedSettings[t]),n),e._raiseChange()},e.getSettings=function(t,n,r){r===void 0&&(r=bK);for(var o={},i=n&&r.scopedSettings[n]||{},a=n&&Ol.scopedSettings[n]||{},s=0,l=t;s<l.length;s++){var u=l[s];o[u]=i[u]||r.settings[u]||a[u]||Ol.settings[u]}return o},e.applyBatchedUpdates=function(t,n){e._suppressUpdates=!0;try{t()}catch{}e._suppressUpdates=!1,n||e._raiseChange()},e.observe=function(t){Em.push(t)},e.unobserve=function(t){Em=Em.filter(function(n){return n!==t})},e._raiseChange=function(){e._suppressUpdates||Em.forEach(function(t){return t()})},e}(),o0=T.createContext({customizations:{inCustomizerContext:!1,settings:{},scopedSettings:{}}});function yK(e,t){e===void 0&&(e={});var n=MF(t)?t:_K(t);return n(e)}function EK(e,t){e===void 0&&(e={});var n=MF(t)?t:TK(t);return n(e)}function MF(e){return typeof e=="function"}function _K(e){return function(t){return e?oe(oe({},t),e):t}}function TK(e){return e===void 0&&(e={}),function(t){var n=oe({},t);for(var r in e)e.hasOwnProperty(r)&&(n[r]=oe(oe({},t[r]),e[r]));return n}}function wK(e,t){var n=(t||{}).customizations,r=n===void 0?{settings:{},scopedSettings:{}}:n;return{customizations:{settings:yK(r.settings,e.settings),scopedSettings:EK(r.scopedSettings,e.scopedSettings),inCustomizerContext:!0}}}var kK=function(e){yi(t,e);function t(){var n=e!==null&&e.apply(this,arguments)||this;return n._onCustomizationChange=function(){return n.forceUpdate()},n}return t.prototype.componentDidMount=function(){$i.observe(this._onCustomizationChange)},t.prototype.componentWillUnmount=function(){$i.unobserve(this._onCustomizationChange)},t.prototype.render=function(){var n=this,r=this.props.contextTransform;return T.createElement(o0.Consumer,null,function(o){var i=wK(n.props,o);return r&&(i=r(i)),T.createElement(o0.Provider,{value:i},n.props.children)})},t}(T.Component);function SK(e,t){for(var n in e)e.hasOwnProperty(n)&&(t[n]=e[n]);return t}function xK(e,t,n){return function(o){var i,a=(i=function(s){yi(l,s);function l(u){var d=s.call(this,u)||this;return d._styleCache={},d._onSettingChanged=d._onSettingChanged.bind(d),d}return l.prototype.componentDidMount=function(){$i.observe(this._onSettingChanged)},l.prototype.componentWillUnmount=function(){$i.unobserve(this._onSettingChanged)},l.prototype.render=function(){var u=this;return T.createElement(o0.Consumer,null,function(d){var h=$i.getSettings(t,e,d.customizations),p=u.props;if(h.styles&&typeof h.styles=="function"&&(h.styles=h.styles(oe(oe({},h),p))),n&&h.styles){if(u._styleCache.default!==h.styles||u._styleCache.component!==p.styles){var m=jc(h.styles,p.styles);u._styleCache.default=h.styles,u._styleCache.component=p.styles,u._styleCache.merged=m}return T.createElement(o,oe({},h,p,{styles:u._styleCache.merged}))}return T.createElement(o,oe({},h,p))})},l.prototype._onSettingChanged=function(){this.forceUpdate()},l}(T.Component),i.displayName="Customized"+e,i);return SK(o,a)}}function CK(e,t){var n=AK(),r=T.useContext(o0).customizations,o=r.inCustomizerContext;return T.useEffect(function(){return o||$i.observe(n),function(){o||$i.unobserve(n)}},[o]),$i.getSettings(e,t,r)}function AK(){var e=T.useState(0),t=e[1];return function(){return t(function(n){return++n})}}function NK(e,t){for(var n in t)t.hasOwnProperty(n)&&(e[n]=UG(e,e[n],t[n]))}var ov="__currentId__",FK="id__",iv=ur()||{};iv[ov]===void 0&&(iv[ov]=0);var ix=!1;function cu(e){if(!ix){var t=Qi.getInstance();t&&t.onReset&&t.onReset(IK),ix=!0}var n=iv[ov]++;return(e===void 0?FK:e)+n}function IK(e){e===void 0&&(e=0),iv[ov]=e}var Xn=function(){for(var e=[],t=0;t<arguments.length;t++)e[t]=arguments[t];for(var n={},r=0,o=e;r<o.length;r++)for(var i=o[r],a=Array.isArray(i)?i:Object.keys(i),s=0,l=a;s<l.length;s++){var u=l[s];n[u]=1}return n},BK=Xn(["onCopy","onCut","onPaste","onCompositionEnd","onCompositionStart","onCompositionUpdate","onFocus","onFocusCapture","onBlur","onBlurCapture","onChange","onInput","onSubmit","onLoad","onError","onKeyDown","onKeyDownCapture","onKeyPress","onKeyUp","onAbort","onCanPlay","onCanPlayThrough","onDurationChange","onEmptied","onEncrypted","onEnded","onLoadedData","onLoadedMetadata","onLoadStart","onPause","onPlay","onPlaying","onProgress","onRateChange","onSeeked","onSeeking","onStalled","onSuspend","onTimeUpdate","onVolumeChange","onWaiting","onClick","onClickCapture","onContextMenu","onDoubleClick","onDrag","onDragEnd","onDragEnter","onDragExit","onDragLeave","onDragOver","onDragStart","onDrop","onMouseDown","onMouseDownCapture","onMouseEnter","onMouseLeave","onMouseMove","onMouseOut","onMouseOver","onMouseUp","onMouseUpCapture","onSelect","onTouchCancel","onTouchEnd","onTouchMove","onTouchStart","onScroll","onWheel","onPointerCancel","onPointerDown","onPointerEnter","onPointerLeave","onPointerMove","onPointerOut","onPointerOver","onPointerUp","onGotPointerCapture","onLostPointerCapture"]),RK=Xn(["accessKey","children","className","contentEditable","dir","draggable","hidden","htmlFor","id","lang","ref","role","style","tabIndex","title","translate","spellCheck","name"]),fr=Xn(RK,BK);Xn(fr,["form"]);var OK=Xn(fr,["height","loop","muted","preload","src","width"]);Xn(OK,["poster"]);Xn(fr,["start"]);Xn(fr,["value"]);var LF=Xn(fr,["download","href","hrefLang","media","rel","target","type"]),Oc=Xn(fr,["autoFocus","disabled","form","formAction","formEncType","formMethod","formNoValidate","formTarget","type","value"]);Xn(Oc,["accept","alt","autoCapitalize","autoComplete","checked","dirname","form","height","inputMode","list","max","maxLength","min","minLength","multiple","pattern","placeholder","readOnly","required","src","step","size","type","value","width"]);Xn(Oc,["autoCapitalize","cols","dirname","form","maxLength","minLength","placeholder","readOnly","required","rows","wrap"]);Xn(Oc,["form","multiple","required"]);Xn(fr,["selected","value"]);Xn(fr,["cellPadding","cellSpacing"]);Xn(fr,["rowSpan","scope"]);Xn(fr,["colSpan","headers","rowSpan","scope"]);Xn(fr,["span"]);Xn(fr,["span"]);Xn(fr,["acceptCharset","action","encType","encType","method","noValidate","target"]);Xn(fr,["allow","allowFullScreen","allowPaymentRequest","allowTransparency","csp","height","importance","referrerPolicy","sandbox","src","srcDoc","width"]);var DK=Xn(fr,["alt","crossOrigin","height","src","srcSet","useMap","width"]),W0=fr;function Zr(e,t,n){for(var r=Array.isArray(t),o={},i=Object.keys(e),a=0,s=i;a<s.length;a++){var l=s[a],u=!r&&t[l]||r&&t.indexOf(l)>=0||l.indexOf("data-")===0||l.indexOf("aria-")===0;u&&(!n||(n==null?void 0:n.indexOf(l))===-1)&&(o[l]=e[l])}return o}function Tb(e){NK(e,{componentDidMount:PK,componentDidUpdate:MK,componentWillUnmount:LK})}function PK(){av(this.props.componentRef,this)}function MK(e){e.componentRef!==this.props.componentRef&&(av(e.componentRef,null),av(this.props.componentRef,this))}function LK(){av(this.props.componentRef,null)}function av(e,t){e&&(typeof e=="object"?e.current=t:typeof e=="function"&&e(t))}var ca,jK=(ca={},ca[qt.up]=1,ca[qt.down]=1,ca[qt.left]=1,ca[qt.right]=1,ca[qt.home]=1,ca[qt.end]=1,ca[qt.tab]=1,ca[qt.pageUp]=1,ca[qt.pageDown]=1,ca);function zK(e){return!!jK[e]}var Go="ms-Fabric--isFocusVisible",ax="ms-Fabric--isFocusHidden";function dd(e,t){var n=t?ur(t):ur();if(n){var r=n.document.body.classList;r.add(e?Go:ax),r.remove(e?ax:Go)}}var sx=new WeakMap;function lx(e,t){var n,r=sx.get(e);return r?n=r+t:n=1,sx.set(e,n),n}function jF(e){T.useEffect(function(){var t,n=ur(e==null?void 0:e.current);if(!(!n||((t=n.FabricConfig)===null||t===void 0?void 0:t.disableFocusRects)===!0)){var r=lx(n,1);return r<=1&&(n.addEventListener("mousedown",ux,!0),n.addEventListener("pointerdown",cx,!0),n.addEventListener("keydown",fx,!0)),function(){var o;!n||((o=n.FabricConfig)===null||o===void 0?void 0:o.disableFocusRects)===!0||(r=lx(n,-1),r===0&&(n.removeEventListener("mousedown",ux,!0),n.removeEventListener("pointerdown",cx,!0),n.removeEventListener("keydown",fx,!0)))}}},[e])}var HK=function(e){return jF(e.rootRef),null};function ux(e){dd(!1,e.target)}function cx(e){e.pointerType!=="mouse"&&dd(!1,e.target)}function fx(e){zK(e.which)&&dd(!0,e.target)}function UK(e){var t=null;try{var n=ur();t=n?n.localStorage.getItem(e):null}catch{}return t}var tc,dx="language";function qK(e){if(e===void 0&&(e="sessionStorage"),tc===void 0){var t=ss(),n=e==="localStorage"?UK(dx):e==="sessionStorage"?FF(dx):void 0;n&&(tc=n),tc===void 0&&t&&(tc=t.documentElement.getAttribute("lang")),tc===void 0&&(tc="en")}return tc}function hx(e){for(var t=[],n=1;n<arguments.length;n++)t[n-1]=arguments[n];for(var r=0,o=t;r<o.length;r++){var i=o[r];zF(e||{},i)}return e}function zF(e,t,n){n===void 0&&(n=[]),n.push(t);for(var r in t)if(t.hasOwnProperty(r)&&r!=="__proto__"&&r!=="constructor"&&r!=="prototype"){var o=t[r];if(typeof o=="object"&&o!==null&&!Array.isArray(o)){var i=n.indexOf(o)>-1;e[r]=i?o:zF(e[r]||{},o,n)}else e[r]=o}return n.pop(),e}var px=function(){return!window||!window.navigator||!window.navigator.userAgent?!1:/iPad|iPhone|iPod/i.test(window.navigator.userAgent)},$K=["TEMPLATE","STYLE","SCRIPT"];function WK(e){var t=ss(e);if(!t)return function(){};for(var n=[];e!==t.body&&e.parentElement;){for(var r=0,o=e.parentElement.children;r<o.length;r++){var i=o[r],a=i.getAttribute("aria-hidden");i!==e&&(a==null?void 0:a.toLowerCase())!=="true"&&$K.indexOf(i.tagName)===-1&&n.push([i,a])}e=e.parentElement}return n.forEach(function(s){var l=s[0];l.setAttribute("aria-hidden","true")}),function(){GK(n),n=[]}}function GK(e){e.forEach(function(t){var n=t[0],r=t[1];r?n.setAttribute("aria-hidden",r):n.removeAttribute("aria-hidden")})}var r2;function mx(e){var t;if(typeof r2>"u"||e){var n=ur(),r=(t=n==null?void 0:n.navigator)===null||t===void 0?void 0:t.userAgent;r2=!!r&&r.indexOf("Macintosh")!==-1}return!!r2}function KK(e){var t=Od(function(n){var r=Od(function(o){return function(i){return n(i,o)}});return function(o,i){return e(o,i?r(i):n)}});return t}var VK=Od(KK);function HF(e,t){return VK(e)(t)}var YK=["theme","styles"];function gl(e,t,n,r,o){r=r||{scope:"",fields:void 0};var i=r.scope,a=r.fields,s=a===void 0?YK:a,l=T.forwardRef(function(d,h){var p=T.useRef(),m=CK(s,i),v=m.styles;m.dir;var _=pl(m,["styles","dir"]),b=n?n(d):void 0,E=p.current&&p.current.__cachedInputs__||[],w=d.styles;if(!p.current||v!==E[1]||w!==E[2]){var k=function(y){return kF(y,t,v,w)};k.__cachedInputs__=[t,v,w],k.__noStyleOverride__=!v&&!w,p.current=k}return T.createElement(e,oe({ref:h},_,b,d,{styles:p.current}))});l.displayName="Styled"+(e.displayName||e.name);var u=o?T.memo(l):l;return l.displayName&&(u.displayName=l.displayName),u}function S4(e,t){for(var n=oe({},t),r=0,o=Object.keys(e);r<o.length;r++){var i=o[r];n[i]===void 0&&(n[i]=e[i])}return n}var QK=function(e){return function(t){for(var n=0,r=e.refs;n<r.length;n++){var o=r[n];typeof o=="function"?o(t):o&&(o.current=t)}}},XK=function(e){var t={refs:[]};return function(){for(var n=[],r=0;r<arguments.length;r++)n[r]=arguments[r];return(!t.resolver||!$G(t.refs,n))&&(t.resolver=QK(t)),t.refs=n,t.resolver}},i0=T.useLayoutEffect,ZK="icons",mi=NF.getValue(ZK,{__options:{disableWarnings:!1,warnOnMissingIcons:!0},__remapped:{}}),o2=Qi.getInstance();o2&&o2.onReset&&o2.onReset(function(){for(var e in mi)mi.hasOwnProperty(e)&&mi[e].subset&&(mi[e].subset.className=void 0)});var sv=function(e){return e.toLowerCase()};function Ar(e,t){var n=oe(oe({},e),{isRegistered:!1,className:void 0}),r=e.icons;t=t?oe(oe({},mi.__options),t):mi.__options;for(var o in r)if(r.hasOwnProperty(o)){var i=r[o],a=sv(o);mi[a]?eV(o):mi[a]={code:i,subset:n}}}function nc(e,t){mi.__remapped[sv(e)]=sv(t)}function JK(e){var t=void 0,n=mi.__options;if(e=e?sv(e):"",e=mi.__remapped[e]||e,e)if(t=mi[e],t){var r=t.subset;r&&r.fontFace&&(r.isRegistered||(SF(r.fontFace),r.isRegistered=!0),r.className||(r.className=Fo(r.style,{fontFamily:r.fontFace.fontFamily,fontWeight:r.fontFace.fontWeight||"normal",fontStyle:r.fontFace.fontStyle||"normal"})))}else!n.disableWarnings&&n.warnOnMissingIcons&&xF('The icon "'+e+'" was used but not registered. See https://github.com/microsoft/fluentui/wiki/Using-icons for more information.');return t}var z1=[],i2=void 0;function eV(e){var t=mi.__options,n=2e3,r=10;t.disableWarnings||(z1.push(e),i2===void 0&&(i2=setTimeout(function(){xF(`Some icons were re-registered. Applications should only call registerIcons for any given icon once. Redefining what an icon is may have unintended consequences. Duplicates include:
 `+z1.slice(0,r).join(", ")+(z1.length>r?" (+ "+(z1.length-r)+" more)":"")),i2=void 0,z1=[]},n)))}function tV(e,t,n,r,o){o===void 0&&(o=!1);var i=oe({primaryButtonBorder:"transparent",errorText:r?"#F1707B":"#a4262c",messageText:r?"#F3F2F1":"#323130",messageLink:r?"#6CB8F6":"#005A9E",messageLinkHovered:r?"#82C7FF":"#004578",infoIcon:r?"#C8C6C4":"#605e5c",errorIcon:r?"#F1707B":"#A80000",blockingIcon:r?"#442726":"#FDE7E9",warningIcon:r?"#C8C6C4":"#797775",severeWarningIcon:r?"#FCE100":"#D83B01",successIcon:r?"#92C353":"#107C10",infoBackground:r?"#323130":"#f3f2f1",errorBackground:r?"#442726":"#FDE7E9",blockingBackground:r?"#442726":"#FDE7E9",warningBackground:r?"#433519":"#FFF4CE",severeWarningBackground:r?"#4F2A0F":"#FED9CC",successBackground:r?"#393D1B":"#DFF6DD",warningHighlight:r?"#fff100":"#ffb900",successText:r?"#92c353":"#107C10"},n),a=UF(e,t,i,r);return nV(a,o)}function UF(e,t,n,r,o){var i={},a=e||{},s=a.white,l=a.black,u=a.themePrimary,d=a.themeDark,h=a.themeDarker,p=a.themeDarkAlt,m=a.themeLighter,v=a.neutralLight,_=a.neutralLighter,b=a.neutralDark,E=a.neutralQuaternary,w=a.neutralQuaternaryAlt,k=a.neutralPrimary,y=a.neutralSecondary,F=a.neutralSecondaryAlt,C=a.neutralTertiary,A=a.neutralTertiaryAlt,P=a.neutralLighterAlt,I=a.accent;return s&&(i.bodyBackground=s,i.bodyFrameBackground=s,i.accentButtonText=s,i.buttonBackground=s,i.primaryButtonText=s,i.primaryButtonTextHovered=s,i.primaryButtonTextPressed=s,i.inputBackground=s,i.inputForegroundChecked=s,i.listBackground=s,i.menuBackground=s,i.cardStandoutBackground=s),l&&(i.bodyTextChecked=l,i.buttonTextCheckedHovered=l),u&&(i.link=u,i.primaryButtonBackground=u,i.inputBackgroundChecked=u,i.inputIcon=u,i.inputFocusBorderAlt=u,i.menuIcon=u,i.menuHeader=u,i.accentButtonBackground=u),d&&(i.primaryButtonBackgroundPressed=d,i.inputBackgroundCheckedHovered=d,i.inputIconHovered=d),h&&(i.linkHovered=h),p&&(i.primaryButtonBackgroundHovered=p),m&&(i.inputPlaceholderBackgroundChecked=m),v&&(i.bodyBackgroundChecked=v,i.bodyFrameDivider=v,i.bodyDivider=v,i.variantBorder=v,i.buttonBackgroundCheckedHovered=v,i.buttonBackgroundPressed=v,i.listItemBackgroundChecked=v,i.listHeaderBackgroundPressed=v,i.menuItemBackgroundPressed=v,i.menuItemBackgroundChecked=v),_&&(i.bodyBackgroundHovered=_,i.buttonBackgroundHovered=_,i.buttonBackgroundDisabled=_,i.buttonBorderDisabled=_,i.primaryButtonBackgroundDisabled=_,i.disabledBackground=_,i.listItemBackgroundHovered=_,i.listHeaderBackgroundHovered=_,i.menuItemBackgroundHovered=_),E&&(i.primaryButtonTextDisabled=E,i.disabledSubtext=E),w&&(i.listItemBackgroundCheckedHovered=w),C&&(i.disabledBodyText=C,i.variantBorderHovered=(n==null?void 0:n.variantBorderHovered)||C,i.buttonTextDisabled=C,i.inputIconDisabled=C,i.disabledText=C),k&&(i.bodyText=k,i.actionLink=k,i.buttonText=k,i.inputBorderHovered=k,i.inputText=k,i.listText=k,i.menuItemText=k),P&&(i.bodyStandoutBackground=P,i.defaultStateBackground=P),b&&(i.actionLinkHovered=b,i.buttonTextHovered=b,i.buttonTextChecked=b,i.buttonTextPressed=b,i.inputTextHovered=b,i.menuItemTextHovered=b),y&&(i.bodySubtext=y,i.focusBorder=y,i.inputBorder=y,i.smallInputBorder=y,i.inputPlaceholderText=y),F&&(i.buttonBorder=F),A&&(i.disabledBodySubtext=A,i.disabledBorder=A,i.buttonBackgroundChecked=A,i.menuDivider=A),I&&(i.accentButtonBackground=I),t!=null&&t.elevation4&&(i.cardShadow=t.elevation4),!r&&(t!=null&&t.elevation8)?i.cardShadowHovered=t.elevation8:i.variantBorderHovered&&(i.cardShadowHovered="0 0 1px "+i.variantBorderHovered),i=oe(oe({},i),n),i}function nV(e,t){var n="";return t===!0&&(n=" /* @deprecated */"),e.listTextColor=e.listText+n,e.menuItemBackgroundChecked+=n,e.warningHighlight+=n,e.warningText=e.messageText+n,e.successText+=n,e}function rV(e,t){var n,r,o;t===void 0&&(t={});var i=hx({},e,t,{semanticColors:UF(t.palette,t.effects,t.semanticColors,t.isInverted===void 0?e.isInverted:t.isInverted)});if(!((n=t.palette)===null||n===void 0)&&n.themePrimary&&!(!((r=t.palette)===null||r===void 0)&&r.accent)&&(i.palette.accent=t.palette.themePrimary),t.defaultFontStyle)for(var a=0,s=Object.keys(i.fonts);a<s.length;a++){var l=s[a];i.fonts[l]=hx(i.fonts[l],t.defaultFontStyle,(o=t==null?void 0:t.fonts)===null||o===void 0?void 0:o[l])}return i}var gx={themeDarker:"#004578",themeDark:"#005a9e",themeDarkAlt:"#106ebe",themePrimary:"#0078d4",themeSecondary:"#2b88d8",themeTertiary:"#71afe5",themeLight:"#c7e0f4",themeLighter:"#deecf9",themeLighterAlt:"#eff6fc",black:"#000000",blackTranslucent40:"rgba(0,0,0,.4)",neutralDark:"#201f1e",neutralPrimary:"#323130",neutralPrimaryAlt:"#3b3a39",neutralSecondary:"#605e5c",neutralSecondaryAlt:"#8a8886",neutralTertiary:"#a19f9d",neutralTertiaryAlt:"#c8c6c4",neutralQuaternary:"#d2d0ce",neutralQuaternaryAlt:"#e1dfdd",neutralLight:"#edebe9",neutralLighter:"#f3f2f1",neutralLighterAlt:"#faf9f8",accent:"#0078d4",white:"#ffffff",whiteTranslucent40:"rgba(255,255,255,.4)",yellowDark:"#d29200",yellow:"#ffb900",yellowLight:"#fff100",orange:"#d83b01",orangeLight:"#ea4300",orangeLighter:"#ff8c00",redDark:"#a4262c",red:"#e81123",magentaDark:"#5c005c",magenta:"#b4009e",magentaLight:"#e3008c",purpleDark:"#32145a",purple:"#5c2d91",purpleLight:"#b4a0ff",blueDark:"#002050",blueMid:"#00188f",blue:"#0078d4",blueLight:"#00bcf2",tealDark:"#004b50",teal:"#008272",tealLight:"#00b294",greenDark:"#004b1c",green:"#107c10",greenLight:"#bad80a"},Xf;(function(e){e.depth0="0 0 0 0 transparent",e.depth4="0 1.6px 3.6px 0 rgba(0, 0, 0, 0.132), 0 0.3px 0.9px 0 rgba(0, 0, 0, 0.108)",e.depth8="0 3.2px 7.2px 0 rgba(0, 0, 0, 0.132), 0 0.6px 1.8px 0 rgba(0, 0, 0, 0.108)",e.depth16="0 6.4px 14.4px 0 rgba(0, 0, 0, 0.132), 0 1.2px 3.6px 0 rgba(0, 0, 0, 0.108)",e.depth64="0 25.6px 57.6px 0 rgba(0, 0, 0, 0.22), 0 4.8px 14.4px 0 rgba(0, 0, 0, 0.18)"})(Xf||(Xf={}));var vx={elevation4:Xf.depth4,elevation8:Xf.depth8,elevation16:Xf.depth16,elevation64:Xf.depth64,roundedCorner2:"2px",roundedCorner4:"4px",roundedCorner6:"6px"},oV={s2:"4px",s1:"8px",m:"16px",l1:"20px",l2:"32px"},vn="cubic-bezier(.1,.9,.2,1)",Ri="cubic-bezier(.1,.25,.75,.9)",_m="0.167s",bx="0.267s",sn="0.367s",yx="0.467s",Br=Ji({from:{opacity:0},to:{opacity:1}}),Rr=Ji({from:{opacity:1},to:{opacity:0,visibility:"hidden"}}),iV=Fu(-10),aV=Fu(-20),sV=Fu(-40),lV=Fu(-400),uV=Fu(10),cV=Fu(20),fV=Fu(40),dV=Fu(400),hV=wb(10),pV=wb(20),mV=wb(-10),gV=wb(-20),vV=Iu(10),bV=Iu(20),yV=Iu(40),EV=Iu(400),_V=Iu(-10),TV=Iu(-20),wV=Iu(-40),kV=Iu(-400),SV=kb(-10),xV=kb(-20),CV=kb(10),AV=kb(20),NV=Ji({from:{transform:"scale3d(.98,.98,1)"},to:{transform:"scale3d(1,1,1)"}}),FV=Ji({from:{transform:"scale3d(1,1,1)"},to:{transform:"scale3d(.98,.98,1)"}}),IV=Ji({from:{transform:"scale3d(1.03,1.03,1)"},to:{transform:"scale3d(1,1,1)"}}),BV=Ji({from:{transform:"scale3d(1,1,1)"},to:{transform:"scale3d(1.03,1.03,1)"}}),RV=Ji({from:{transform:"rotateZ(0deg)"},to:{transform:"rotateZ(90deg)"}}),OV=Ji({from:{transform:"rotateZ(0deg)"},to:{transform:"rotateZ(-90deg)"}}),DV={slideRightIn10:Ct(Br+","+iV,sn,vn),slideRightIn20:Ct(Br+","+aV,sn,vn),slideRightIn40:Ct(Br+","+sV,sn,vn),slideRightIn400:Ct(Br+","+lV,sn,vn),slideLeftIn10:Ct(Br+","+uV,sn,vn),slideLeftIn20:Ct(Br+","+cV,sn,vn),slideLeftIn40:Ct(Br+","+fV,sn,vn),slideLeftIn400:Ct(Br+","+dV,sn,vn),slideUpIn10:Ct(Br+","+hV,sn,vn),slideUpIn20:Ct(Br+","+pV,sn,vn),slideDownIn10:Ct(Br+","+mV,sn,vn),slideDownIn20:Ct(Br+","+gV,sn,vn),slideRightOut10:Ct(Rr+","+vV,sn,vn),slideRightOut20:Ct(Rr+","+bV,sn,vn),slideRightOut40:Ct(Rr+","+yV,sn,vn),slideRightOut400:Ct(Rr+","+EV,sn,vn),slideLeftOut10:Ct(Rr+","+_V,sn,vn),slideLeftOut20:Ct(Rr+","+TV,sn,vn),slideLeftOut40:Ct(Rr+","+wV,sn,vn),slideLeftOut400:Ct(Rr+","+kV,sn,vn),slideUpOut10:Ct(Rr+","+SV,sn,vn),slideUpOut20:Ct(Rr+","+xV,sn,vn),slideDownOut10:Ct(Rr+","+CV,sn,vn),slideDownOut20:Ct(Rr+","+AV,sn,vn),scaleUpIn100:Ct(Br+","+NV,sn,vn),scaleDownIn100:Ct(Br+","+IV,sn,vn),scaleUpOut103:Ct(Rr+","+BV,_m,Ri),scaleDownOut98:Ct(Rr+","+FV,_m,Ri),fadeIn100:Ct(Br,_m,Ri),fadeIn200:Ct(Br,bx,Ri),fadeIn400:Ct(Br,sn,Ri),fadeIn500:Ct(Br,yx,Ri),fadeOut100:Ct(Rr,_m,Ri),fadeOut200:Ct(Rr,bx,Ri),fadeOut400:Ct(Rr,sn,Ri),fadeOut500:Ct(Rr,yx,Ri),rotate90deg:Ct(RV,"0.1s",Ri),rotateN90deg:Ct(OV,"0.1s",Ri)};function Ct(e,t,n){return{animationName:e,animationDuration:t,animationTimingFunction:n,animationFillMode:"both"}}function Fu(e){return Ji({from:{transform:"translate3d("+e+"px,0,0)",pointerEvents:"none"},to:{transform:"translate3d(0,0,0)",pointerEvents:"auto"}})}function wb(e){return Ji({from:{transform:"translate3d(0,"+e+"px,0)",pointerEvents:"none"},to:{transform:"translate3d(0,0,0)",pointerEvents:"auto"}})}function Iu(e){return Ji({from:{transform:"translate3d(0,0,0)"},to:{transform:"translate3d("+e+"px,0,0)"}})}function kb(e){return Ji({from:{transform:"translate3d(0,0,0)"},to:{transform:"translate3d(0,"+e+"px,0)"}})}var zn;(function(e){e.Arabic="Segoe UI Web (Arabic)",e.Cyrillic="Segoe UI Web (Cyrillic)",e.EastEuropean="Segoe UI Web (East European)",e.Greek="Segoe UI Web (Greek)",e.Hebrew="Segoe UI Web (Hebrew)",e.Thai="Leelawadee UI Web",e.Vietnamese="Segoe UI Web (Vietnamese)",e.WestEuropean="Segoe UI Web (West European)",e.Selawik="Selawik Web",e.Armenian="Segoe UI Web (Armenian)",e.Georgian="Segoe UI Web (Georgian)"})(zn||(zn={}));var rn;(function(e){e.Arabic="'"+zn.Arabic+"'",e.ChineseSimplified="'Microsoft Yahei UI', Verdana, Simsun",e.ChineseTraditional="'Microsoft Jhenghei UI', Pmingliu",e.Cyrillic="'"+zn.Cyrillic+"'",e.EastEuropean="'"+zn.EastEuropean+"'",e.Greek="'"+zn.Greek+"'",e.Hebrew="'"+zn.Hebrew+"'",e.Hindi="'Nirmala UI'",e.Japanese="'Yu Gothic UI', 'Meiryo UI', Meiryo, 'MS Pgothic', Osaka",e.Korean="'Malgun Gothic', Gulim",e.Selawik="'"+zn.Selawik+"'",e.Thai="'Leelawadee UI Web', 'Kmer UI'",e.Vietnamese="'"+zn.Vietnamese+"'",e.WestEuropean="'"+zn.WestEuropean+"'",e.Armenian="'"+zn.Armenian+"'",e.Georgian="'"+zn.Georgian+"'"})(rn||(rn={}));var So;(function(e){e.size10="10px",e.size12="12px",e.size14="14px",e.size16="16px",e.size18="18px",e.size20="20px",e.size24="24px",e.size28="28px",e.size32="32px",e.size42="42px",e.size68="68px",e.mini="10px",e.xSmall="10px",e.small="12px",e.smallPlus="12px",e.medium="14px",e.mediumPlus="16px",e.icon="16px",e.large="18px",e.xLarge="20px",e.xLargePlus="24px",e.xxLarge="28px",e.xxLargePlus="32px",e.superLarge="42px",e.mega="68px"})(So||(So={}));var Rn;(function(e){e.light=100,e.semilight=300,e.regular=400,e.semibold=600,e.bold=700})(Rn||(Rn={}));var Kl;(function(e){e.xSmall="10px",e.small="12px",e.medium="16px",e.large="20px"})(Kl||(Kl={}));var PV="'Segoe UI', -apple-system, BlinkMacSystemFont, 'Roboto', 'Helvetica Neue', sans-serif",MV="'Segoe UI', '"+zn.WestEuropean+"'",a2={ar:rn.Arabic,bg:rn.Cyrillic,cs:rn.EastEuropean,el:rn.Greek,et:rn.EastEuropean,he:rn.Hebrew,hi:rn.Hindi,hr:rn.EastEuropean,hu:rn.EastEuropean,ja:rn.Japanese,kk:rn.EastEuropean,ko:rn.Korean,lt:rn.EastEuropean,lv:rn.EastEuropean,pl:rn.EastEuropean,ru:rn.Cyrillic,sk:rn.EastEuropean,"sr-latn":rn.EastEuropean,th:rn.Thai,tr:rn.EastEuropean,uk:rn.Cyrillic,vi:rn.Vietnamese,"zh-hans":rn.ChineseSimplified,"zh-hant":rn.ChineseTraditional,hy:rn.Armenian,ka:rn.Georgian};function LV(e){return e+", "+PV}function jV(e){for(var t in a2)if(a2.hasOwnProperty(t)&&e&&t.indexOf(e)===0)return a2[t];return MV}function fi(e,t,n){return{fontFamily:n,MozOsxFontSmoothing:"grayscale",WebkitFontSmoothing:"antialiased",fontSize:e,fontWeight:t}}function zV(e){var t=jV(e),n=LV(t),r={tiny:fi(So.mini,Rn.regular,n),xSmall:fi(So.xSmall,Rn.regular,n),small:fi(So.small,Rn.regular,n),smallPlus:fi(So.smallPlus,Rn.regular,n),medium:fi(So.medium,Rn.regular,n),mediumPlus:fi(So.mediumPlus,Rn.regular,n),large:fi(So.large,Rn.regular,n),xLarge:fi(So.xLarge,Rn.semibold,n),xLargePlus:fi(So.xLargePlus,Rn.semibold,n),xxLarge:fi(So.xxLarge,Rn.semibold,n),xxLargePlus:fi(So.xxLargePlus,Rn.semibold,n),superLarge:fi(So.superLarge,Rn.semibold,n),mega:fi(So.mega,Rn.semibold,n)};return r}var HV="https://static2.sharepointonline.com/files/fabric/assets",UV=zV(qK());function mc(e,t,n,r){e="'"+e+"'";var o=r!==void 0?"local('"+r+"'),":"";SF({fontFamily:e,src:o+("url('"+t+".woff2') format('woff2'),")+("url('"+t+".woff') format('woff')"),fontWeight:n,fontStyle:"normal",fontDisplay:"swap"})}function fa(e,t,n,r,o){r===void 0&&(r="segoeui");var i=e+"/"+n+"/"+r;mc(t,i+"-light",Rn.light,o&&o+" Light"),mc(t,i+"-semilight",Rn.semilight,o&&o+" SemiLight"),mc(t,i+"-regular",Rn.regular,o),mc(t,i+"-semibold",Rn.semibold,o&&o+" SemiBold"),mc(t,i+"-bold",Rn.bold,o&&o+" Bold")}function qV(e){if(e){var t=e+"/fonts";fa(t,zn.Thai,"leelawadeeui-thai","leelawadeeui"),fa(t,zn.Arabic,"segoeui-arabic"),fa(t,zn.Cyrillic,"segoeui-cyrillic"),fa(t,zn.EastEuropean,"segoeui-easteuropean"),fa(t,zn.Greek,"segoeui-greek"),fa(t,zn.Hebrew,"segoeui-hebrew"),fa(t,zn.Vietnamese,"segoeui-vietnamese"),fa(t,zn.WestEuropean,"segoeui-westeuropean","segoeui","Segoe UI"),fa(t,rn.Selawik,"selawik","selawik"),fa(t,zn.Armenian,"segoeui-armenian"),fa(t,zn.Georgian,"segoeui-georgian"),mc("Leelawadee UI Web",t+"/leelawadeeui-thai/leelawadeeui-semilight",Rn.light),mc("Leelawadee UI Web",t+"/leelawadeeui-thai/leelawadeeui-bold",Rn.semibold)}}function $V(){var e,t,n=(e=ur())===null||e===void 0?void 0:e.FabricConfig;return(t=n==null?void 0:n.fontBaseUrl)!==null&&t!==void 0?t:HV}qV($V());function Sb(e,t){e===void 0&&(e={}),t===void 0&&(t=!1);var n=!!e.isInverted,r={palette:gx,effects:vx,fonts:UV,spacing:oV,isInverted:n,disableGlobalClassNames:!1,semanticColors:tV(gx,vx,void 0,n,t),rtl:void 0};return rV(r,e)}var Ao="@media screen and (-ms-high-contrast: active), (forced-colors: active)",WV=640,qF=WV-1;function $F(e,t){var n=typeof e=="number"?" and (min-width: "+e+"px)":"",r=typeof t=="number"?" and (max-width: "+t+"px)":"";return"@media only screen"+n+r}function WF(){return{forcedColorAdjust:"none",MsHighContrastAdjust:"none"}}var Dd;(function(e){e.Nav=1,e.ScrollablePane=1,e.FocusStyle=1,e.Coachmark=1e3,e.Layer=1e6,e.KeytipLayer=1000001})(Dd||(Dd={}));function Pd(e,t,n,r,o,i,a){return typeof t=="number"||!t?Ex(e,{inset:t,position:n,highContrastStyle:r,borderColor:o,outlineColor:i,isFocusedOnly:a}):Ex(e,t)}function Ex(e,t){var n,r;t===void 0&&(t={});var o=t.inset,i=o===void 0?0:o,a=t.width,s=a===void 0?1:a,l=t.position,u=l===void 0?"relative":l,d=t.highContrastStyle,h=t.borderColor,p=h===void 0?e.palette.white:h,m=t.outlineColor,v=m===void 0?e.palette.neutralSecondary:m,_=t.isFocusedOnly,b=_===void 0?!0:_;return{outline:"transparent",position:u,selectors:(n={"::-moz-focus-inner":{border:"0"}},n["."+Go+" &"+(b?":focus":"")+":after"]={content:'""',position:"absolute",left:i+1,top:i+1,bottom:i+1,right:i+1,border:s+"px solid "+p,outline:s+"px solid "+v,zIndex:Dd.FocusStyle,selectors:(r={},r[Ao]=d,r)},n)}}function GV(){return{selectors:{"&::-moz-focus-inner":{border:0},"&":{outline:"transparent"}}}}var GF={position:"absolute",width:1,height:1,margin:-1,padding:0,border:0,overflow:"hidden",whiteSpace:"nowrap"},KV=Cr(function(e,t){var n=Qi.getInstance();return t?Object.keys(e).reduce(function(r,o){return r[o]=n.getClassName(e[o]),r},{}):e});function hs(e,t,n){return KV(e,n!==void 0?n:t.disableGlobalClassNames)}var lv=globalThis&&globalThis.__assign||function(){return lv=Object.assign||function(e){for(var t,n=1,r=arguments.length;n<r;n++){t=arguments[n];for(var o in t)Object.prototype.hasOwnProperty.call(t,o)&&(e[o]=t[o])}return e},lv.apply(this,arguments)},Ch=typeof window>"u"?global:window,_x=Ch&&Ch.CSPSettings&&Ch.CSPSettings.nonce,gi=VV();function VV(){var e=Ch.__themeState__||{theme:void 0,lastStyleElement:void 0,registeredStyles:[]};return e.runState||(e=lv({},e,{perf:{count:0,duration:0},runState:{flushTimer:0,mode:0,buffer:[]}})),e.registeredThemableStyles||(e=lv({},e,{registeredThemableStyles:[]})),Ch.__themeState__=e,e}function YV(e,t){gi.loadStyles?gi.loadStyles(VF(e).styleString,e):ZV(e)}function KF(e){gi.theme=e,XV()}function QV(e){e===void 0&&(e=3),(e===3||e===2)&&(Tx(gi.registeredStyles),gi.registeredStyles=[]),(e===3||e===1)&&(Tx(gi.registeredThemableStyles),gi.registeredThemableStyles=[])}function Tx(e){e.forEach(function(t){var n=t&&t.styleElement;n&&n.parentElement&&n.parentElement.removeChild(n)})}function XV(){if(gi.theme){for(var e=[],t=0,n=gi.registeredThemableStyles;t<n.length;t++){var r=n[t];e.push(r.themableStyle)}e.length>0&&(QV(1),YV([].concat.apply([],e)))}}function VF(e){var t=gi.theme,n=!1,r=(e||[]).map(function(o){var i=o.theme;if(i){n=!0;var a=t?t[i]:void 0,s=o.defaultValue||"inherit";return t&&!a&&console&&!(i in t)&&typeof DEBUG<"u"&&DEBUG&&console.warn('Theming value not provided for "'+i+'". Falling back to "'+s+'".'),a||s}else return o.rawString});return{styleString:r.join(""),themable:n}}function ZV(e){if(!(typeof document>"u")){var t=document.getElementsByTagName("head")[0],n=document.createElement("style"),r=VF(e),o=r.styleString,i=r.themable;n.setAttribute("data-load-themed-styles","true"),_x&&n.setAttribute("nonce",_x),n.appendChild(document.createTextNode(o)),gi.perf.count++,t.appendChild(n);var a=document.createEvent("HTMLEvents");a.initEvent("styleinsert",!0,!1),a.args={newStyle:n},document.dispatchEvent(a);var s={styleElement:n,themableStyle:e};i?gi.registeredThemableStyles.push(s):gi.registeredStyles.push(s)}}var La=Sb({}),JV=[],_E="theme";function YF(){var e,t,n,r=ur();!((t=r==null?void 0:r.FabricConfig)===null||t===void 0)&&t.legacyTheme?eY(r.FabricConfig.legacyTheme):$i.getSettings([_E]).theme||(!((n=r==null?void 0:r.FabricConfig)===null||n===void 0)&&n.theme&&(La=Sb(r.FabricConfig.theme)),$i.applySettings((e={},e[_E]=La,e)))}YF();function eY(e,t){var n;return t===void 0&&(t=!1),La=Sb(e,t),KF(oe(oe(oe(oe({},La.palette),La.semanticColors),La.effects),tY(La))),$i.applySettings((n={},n[_E]=La,n)),JV.forEach(function(r){try{r(La)}catch{}}),La}function tY(e){for(var t={},n=0,r=Object.keys(e.fonts);n<r.length;n++)for(var o=r[n],i=e.fonts[o],a=0,s=Object.keys(i);a<s.length;a++){var l=s[a],u=o+l.charAt(0).toUpperCase()+l.slice(1),d=i[l];l==="fontSize"&&typeof d=="number"&&(d=d+"px"),t[u]=d}return t}var bc=DG(DV);yb("@fluentui/style-utilities","8.6.0");YF();var hr={topLeftEdge:0,topCenter:1,topRightEdge:2,topAutoEdge:3,bottomLeftEdge:4,bottomCenter:5,bottomRightEdge:6,bottomAutoEdge:7,leftTopEdge:8,leftCenter:9,leftBottomEdge:10,rightTopEdge:11,rightCenter:12,rightBottomEdge:13},Je;(function(e){e[e.top=1]="top",e[e.bottom=-1]="bottom",e[e.left=2]="left",e[e.right=-2]="right"})(Je||(Je={}));var wx;(function(e){e[e.top=0]="top",e[e.bottom=1]="bottom",e[e.start=2]="start",e[e.end=3]="end"})(wx||(wx={}));var oo;function Ho(e,t,n){return{targetEdge:e,alignmentEdge:t,isAuto:n}}var kx=(oo={},oo[hr.topLeftEdge]=Ho(Je.top,Je.left),oo[hr.topCenter]=Ho(Je.top),oo[hr.topRightEdge]=Ho(Je.top,Je.right),oo[hr.topAutoEdge]=Ho(Je.top,void 0,!0),oo[hr.bottomLeftEdge]=Ho(Je.bottom,Je.left),oo[hr.bottomCenter]=Ho(Je.bottom),oo[hr.bottomRightEdge]=Ho(Je.bottom,Je.right),oo[hr.bottomAutoEdge]=Ho(Je.bottom,void 0,!0),oo[hr.leftTopEdge]=Ho(Je.left,Je.top),oo[hr.leftCenter]=Ho(Je.left),oo[hr.leftBottomEdge]=Ho(Je.left,Je.bottom),oo[hr.rightTopEdge]=Ho(Je.right,Je.top),oo[hr.rightCenter]=Ho(Je.right),oo[hr.rightBottomEdge]=Ho(Je.right,Je.bottom),oo);function x4(e,t){return!(e.top<t.top||e.bottom>t.bottom||e.left<t.left||e.right>t.right)}function uv(e,t){var n=[];return e.top<t.top&&n.push(Je.top),e.bottom>t.bottom&&n.push(Je.bottom),e.left<t.left&&n.push(Je.left),e.right>t.right&&n.push(Je.right),n}function po(e,t){return e[Je[t]]}function Sx(e,t,n){return e[Je[t]]=n,e}function a0(e,t){var n=Xd(t);return(po(e,n.positiveEdge)+po(e,n.negativeEdge))/2}function xb(e,t){return e>0?t:t*-1}function TE(e,t){return xb(e,po(t,e))}function Qs(e,t,n){var r=po(e,n)-po(t,n);return xb(n,r)}function Md(e,t,n,r){r===void 0&&(r=!0);var o=po(e,t)-n,i=Sx(e,t,n);return r&&(i=Sx(e,t*-1,po(e,t*-1)-o)),i}function s0(e,t,n,r){return r===void 0&&(r=0),Md(e,n,po(t,n)+xb(n,r))}function nY(e,t,n,r){r===void 0&&(r=0);var o=n*-1,i=xb(o,r);return Md(e,n*-1,po(t,n)+i)}function cv(e,t,n){var r=TE(n,e);return r>TE(n,t)}function rY(e,t){for(var n=uv(e,t),r=0,o=0,i=n;o<i.length;o++){var a=i[o];r+=Math.pow(Qs(e,t,a),2)}return r}function oY(e,t,n,r,o){o===void 0&&(o=0);var i=[Je.left,Je.right,Je.bottom,Je.top];ls()&&(i[0]*=-1,i[1]*=-1);for(var a=e,s=r.targetEdge,l=r.alignmentEdge,u,d=s,h=l,p=0;p<4;p++){if(cv(a,n,s))return{elementRectangle:a,targetEdge:s,alignmentEdge:l};var m=rY(a,n);(!u||m<u)&&(u=m,d=s,h=l),i.splice(i.indexOf(s),1),i.length>0&&(i.indexOf(s*-1)>-1?s=s*-1:(l=s,s=i.slice(-1)[0]),a=fv(e,t,{targetEdge:s,alignmentEdge:l},o))}return a=fv(e,t,{targetEdge:d,alignmentEdge:h},o),{elementRectangle:a,targetEdge:d,alignmentEdge:h}}function iY(e,t,n,r){var o=e.alignmentEdge,i=e.targetEdge,a=e.elementRectangle,s=o*-1,l=fv(a,t,{targetEdge:i,alignmentEdge:s},n,r);return{elementRectangle:l,targetEdge:i,alignmentEdge:s}}function aY(e,t,n,r,o,i,a){o===void 0&&(o=0);var s=r.alignmentEdge,l=r.alignTargetEdge,u={elementRectangle:e,targetEdge:r.targetEdge,alignmentEdge:s};!i&&!a&&(u=oY(e,t,n,r,o));var d=uv(u.elementRectangle,n),h=i?-u.targetEdge:void 0;if(d.length>0)if(l)if(u.alignmentEdge&&d.indexOf(u.alignmentEdge*-1)>-1){var p=iY(u,t,o,a);if(x4(p.elementRectangle,n))return p;u=s2(uv(p.elementRectangle,n),u,n,h)}else u=s2(d,u,n,h);else u=s2(d,u,n,h);return u}function s2(e,t,n,r){for(var o=0,i=e;o<i.length;o++){var a=i[o],s=void 0;if(r&&r===a*-1)s=Md(t.elementRectangle,a,po(n,a),!1),t.forcedInBounds=!0;else{s=s0(t.elementRectangle,n,a);var l=cv(s,n,a*-1);l||(s=Md(s,a*-1,po(n,a*-1),!1),t.forcedInBounds=!0)}t.elementRectangle=s}return t}function QF(e,t,n){var r=Xd(t).positiveEdge,o=a0(e,t),i=o-po(e,r);return Md(e,r,n-i)}function fv(e,t,n,r,o){r===void 0&&(r=0);var i=new ts(e.left,e.right,e.top,e.bottom),a=n.alignmentEdge,s=n.targetEdge,l=o?s:s*-1;if(i=o?s0(i,t,s,r):nY(i,t,s,r),a)i=s0(i,t,a);else{var u=a0(t,s);i=QF(i,l,u)}return i}function Xd(e){return e===Je.top||e===Je.bottom?{positiveEdge:Je.left,negativeEdge:Je.right}:{positiveEdge:Je.top,negativeEdge:Je.bottom}}function XF(e,t,n){return n&&Math.abs(Qs(e,n,t))>Math.abs(Qs(e,n,t*-1))?t*-1:t}function sY(e,t,n){return n!==void 0&&po(e,t)===po(n,t)}function lY(e,t,n,r,o,i,a,s){var l={},u=C4(t),d=i?n:n*-1,h=o||Xd(n).positiveEdge;return(!a||sY(e,TY(h),r))&&(h=XF(e,h,r)),l[Je[d]]=Qs(e,u,d),l[Je[h]]=Qs(e,u,h),s&&(l[Je[d*-1]]=Qs(e,u,d*-1),l[Je[h*-1]]=Qs(e,u,h*-1)),l}function uY(e){return Math.sqrt(e*e*2)}function cY(e,t,n){if(e===void 0&&(e=hr.bottomAutoEdge),n)return{alignmentEdge:n.alignmentEdge,isAuto:n.isAuto,targetEdge:n.targetEdge};var r=oe({},kx[e]);return ls()?(r.alignmentEdge&&r.alignmentEdge%2===0&&(r.alignmentEdge=r.alignmentEdge*-1),t!==void 0?kx[t]:r):r}function fY(e,t,n,r,o){return e.isAuto&&(e.alignmentEdge=ZF(e.targetEdge,t,n)),e.alignTargetEdge=o,e}function ZF(e,t,n){var r=a0(t,e),o=a0(n,e),i=Xd(e),a=i.positiveEdge,s=i.negativeEdge;return r<=o?a:s}function dY(e,t,n,r,o,i,a){var s=fv(e,t,r,o,a);return x4(s,n)?{elementRectangle:s,targetEdge:r.targetEdge,alignmentEdge:r.alignmentEdge}:aY(s,t,n,r,o,i,a)}function hY(e,t,n){var r=e.targetEdge*-1,o=new ts(0,e.elementRectangle.width,0,e.elementRectangle.height),i={},a=XF(e.elementRectangle,e.alignmentEdge?e.alignmentEdge:Xd(r).positiveEdge,n),s=Qs(e.elementRectangle,e.targetRectangle,r),l=s>Math.abs(po(t,r));return i[Je[r]]=po(t,r),i[Je[a]]=Qs(t,o,a),{elementPosition:oe({},i),closestEdge:ZF(e.targetEdge,t,o),targetEdge:r,hideBeak:!l}}function pY(e,t){var n=t.targetRectangle,r=Xd(t.targetEdge),o=r.positiveEdge,i=r.negativeEdge,a=a0(n,t.targetEdge),s=new ts(e/2,t.elementRectangle.width-e/2,e/2,t.elementRectangle.height-e/2),l=new ts(0,e,0,e);return l=Md(l,t.targetEdge*-1,-e/2),l=QF(l,t.targetEdge*-1,a-TE(o,t.elementRectangle)),cv(l,s,o)?cv(l,s,i)||(l=s0(l,s,i)):l=s0(l,s,o),l}function C4(e){var t=e.getBoundingClientRect();return new ts(t.left,t.right,t.top,t.bottom)}function mY(e){return new ts(e.left,e.right,e.top,e.bottom)}function gY(e,t){var n;if(t){if(t.preventDefault){var r=t;n=new ts(r.clientX,r.clientX,r.clientY,r.clientY)}else if(t.getBoundingClientRect)n=C4(t);else{var o=t,i=o.left||o.x,a=o.top||o.y,s=o.right||i,l=o.bottom||a;n=new ts(i,s,a,l)}if(!x4(n,e))for(var u=uv(n,e),d=0,h=u;d<h.length;d++){var p=h[d];n[Je[p]]=e[Je[p]]}}else n=new ts(0,0,0,0);return n}function vY(e,t,n,r){var o=e.gapSpace?e.gapSpace:0,i=gY(n,e.target),a=fY(cY(e.directionalHint,e.directionalHintForRTL,r),i,n,e.coverTarget,e.alignTargetEdge),s=dY(C4(t),i,n,a,o,e.directionalHintFixed,e.coverTarget);return oe(oe({},s),{targetRectangle:i})}function bY(e,t,n,r,o){var i=lY(e.elementRectangle,t,e.targetEdge,n,e.alignmentEdge,r,o,e.forcedInBounds);return{elementPosition:i,targetEdge:e.targetEdge,alignmentEdge:e.alignmentEdge}}function JF(e,t,n,r,o){var i=e.isBeakVisible&&e.beakWidth||0,a=uY(i)/2+(e.gapSpace?e.gapSpace:0),s=e;s.gapSpace=a;var l=e.bounds?mY(e.bounds):new ts(0,window.innerWidth-MG(),0,window.innerHeight),u=vY(s,n,l,r),d=pY(i,u),h=hY(u,d,l);return oe(oe({},bY(u,t,l,e.coverTarget,o)),{beakPosition:h})}function yY(e,t,n,r){return JF(e,t,n,r,!0)}function EY(e,t,n,r){return JF(e,t,n,r)}function _Y(e,t,n,r){return yY(e,t,n,r)}function TY(e){return e*-1}function wY(e,t){var n=void 0;if(t.getWindowSegments&&(n=t.getWindowSegments()),n===void 0||n.length<=1)return{top:0,left:0,right:t.innerWidth,bottom:t.innerHeight,width:t.innerWidth,height:t.innerHeight};var r=0,o=0;if(e!==null&&e.getBoundingClientRect){var i=e.getBoundingClientRect();r=(i.left+i.right)/2,o=(i.top+i.bottom)/2}else e!==null&&(r=e.left||e.x,o=e.top||e.y);for(var a={top:0,left:0,right:0,bottom:0,width:0,height:0},s=0,l=n;s<l.length;s++){var u=l[s];r&&u.left<=r&&u.right>=r&&o&&u.top<=o&&u.bottom>=o&&(a={top:u.top,left:u.left,right:u.right,bottom:u.bottom,width:u.width,height:u.height})}return a}function kY(e,t){return wY(e,t)}function A4(e){var t=T.useRef();return t.current===void 0&&(t.current={value:typeof e=="function"?e():e}),t.current.value}function Zd(){var e=A4(function(){return new _b});return T.useEffect(function(){return function(){return e.dispose()}},[e]),e}function eI(e,t){var n=T.useRef(t);return n.current||(n.current=cu(e)),n.current}function G0(){for(var e=[],t=0;t<arguments.length;t++)e[t]=arguments[t];var n=T.useCallback(function(r){n.current=r;for(var o=0,i=e;o<i.length;o++){var a=i[o];typeof a=="function"?a(r):a&&(a.current=r)}},Yi([],e));return n}function l0(e,t,n,r){var o=T.useRef(n);o.current=n,T.useEffect(function(){var i=e&&"current"in e?e.current:e;if(i){var a=zf(i,t,function(s){return o.current(s)},r);return a}},[e,t,r])}function tI(e){var t=T.useRef();return T.useEffect(function(){t.current=e}),t.current}var nI=T.createContext({window:typeof window=="object"?window:void 0}),N4=function(){return T.useContext(nI).window},SY=function(){var e;return(e=T.useContext(nI).window)===null||e===void 0?void 0:e.document};function rI(e,t){var n=T.useRef(),r=T.useRef(null),o=N4();if(!e||e!==n.current||typeof e=="string"){var i=t==null?void 0:t.current;if(e)if(typeof e=="string"){var a=ss(i);r.current=a?a.querySelector(e):null}else"stopPropagation"in e||"getBoundingClientRect"in e?r.current=e:"current"in e?r.current=e.current:r.current=e;n.current=e}return[r,o]}function xY(e,t){var n=Zd(),r=T.useState(!1),o=r[0],i=r[1];return T.useEffect(function(){return n.requestAnimationFrame(function(){var a;if(!(e.style&&e.style.overflowY)){var s=!1;if(t&&t.current&&(!((a=t.current)===null||a===void 0)&&a.firstElementChild)){var l=t.current.clientHeight,u=t.current.firstElementChild.clientHeight;l>0&&u>l&&(s=u-l>1)}o!==s&&i(s)}}),function(){return n.dispose()}}),o}function CY(e){var t=e.originalElement,n=e.containsFocus;t&&n&&t!==ur()&&setTimeout(function(){var r;(r=t.focus)===null||r===void 0||r.call(t)},0)}function AY(e,t){var n=e.onRestoreFocus,r=n===void 0?CY:n,o=T.useRef(),i=T.useRef(!1);T.useEffect(function(){return o.current=ss().activeElement,aK(t.current)&&(i.current=!0),function(){var a;r==null||r({originalElement:o.current,containsFocus:i.current,documentContainsFocus:((a=ss())===null||a===void 0?void 0:a.hasFocus())||!1}),o.current=void 0}},[]),l0(t,"focus",T.useCallback(function(){i.current=!0},[]),!0),l0(t,"blur",T.useCallback(function(a){t.current&&a.relatedTarget&&!t.current.contains(a.relatedTarget)&&(i.current=!1)},[]),!0)}function NY(e,t){var n=String(e["aria-modal"]).toLowerCase()==="true"&&e.enableAriaHiddenSiblings;T.useEffect(function(){if(n&&t.current){var r=WK(t.current);return r}},[t,n])}var oI=T.forwardRef(function(e,t){var n=S4({shouldRestoreFocus:!0,enableAriaHiddenSiblings:!0},e),r=T.useRef(),o=G0(r,t);NY(n,r),AY(n,r);var i=n.role,a=n.className,s=n.ariaLabel,l=n.ariaLabelledBy,u=n.ariaDescribedBy,d=n.style,h=n.children,p=n.onDismiss,m=xY(n,r),v=T.useCallback(function(b){switch(b.which){case qt.escape:p&&(p(b),b.preventDefault(),b.stopPropagation());break}},[p]),_=N4();return l0(_,"keydown",v),T.createElement("div",oe({ref:o},Zr(n,W0),{className:a,role:i,"aria-label":s,"aria-labelledby":l,"aria-describedby":u,onKeyDown:v,style:oe({overflowY:m?"scroll":void 0,outline:"none"},d)}),h)});oI.displayName="Popup";var yf,FY="CalloutContentBase",IY=(yf={},yf[Je.top]=bc.slideUpIn10,yf[Je.bottom]=bc.slideDownIn10,yf[Je.left]=bc.slideLeftIn10,yf[Je.right]=bc.slideRightIn10,yf),xx={top:0,left:0},BY={opacity:0,filter:"opacity(0)",pointerEvents:"none"},RY=["role","aria-roledescription"],iI={preventDismissOnLostFocus:!1,preventDismissOnScroll:!1,preventDismissOnResize:!1,isBeakVisible:!0,beakWidth:16,gapSpace:0,minPagePadding:8,directionalHint:hr.bottomAutoEdge},OY=ml({disableCaching:!0});function DY(e,t,n){var r=e.bounds,o=e.minPagePadding,i=o===void 0?iI.minPagePadding:o,a=e.target,s=T.useState(!1),l=s[0],u=s[1],d=T.useRef(),h=T.useCallback(function(){if(!d.current||l){var m=typeof r=="function"?n?r(a,n):void 0:r;!m&&n&&(m=kY(t.current,n),m={top:m.top+i,left:m.left+i,right:m.right-i,bottom:m.bottom-i,width:m.width-i*2,height:m.height-i*2}),d.current=m,l&&u(!1)}return d.current},[r,i,a,t,n,l]),p=Zd();return l0(n,"resize",p.debounce(function(){u(!0)},500,{leading:!0})),h}function PY(e,t,n){var r,o=e.calloutMaxHeight,i=e.finalHeight,a=e.directionalHint,s=e.directionalHintFixed,l=e.hidden,u=T.useState(),d=u[0],h=u[1],p=(r=n==null?void 0:n.elementPosition)!==null&&r!==void 0?r:{},m=p.top,v=p.bottom;return T.useEffect(function(){var _,b=(_=t())!==null&&_!==void 0?_:{},E=b.top,w=b.bottom;!o&&!l?typeof m=="number"&&w?h(w-m):typeof v=="number"&&typeof E=="number"&&w&&h(w-E-v):h(o||void 0)},[v,o,i,a,s,t,l,n,m]),d}function MY(e,t,n,r,o){var i=T.useState(),a=i[0],s=i[1],l=T.useRef(0),u=T.useRef(),d=Zd(),h=e.hidden,p=e.target,m=e.finalHeight,v=e.calloutMaxHeight,_=e.onPositioned,b=e.directionalHint;return T.useEffect(function(){if(h)s(void 0),l.current=0;else{var E=d.requestAnimationFrame(function(){var w,k;if(t.current&&n){var y=oe(oe({},e),{target:r.current,bounds:o()}),F=n.cloneNode(!0);F.style.maxHeight=v?""+v:"",F.style.visibility="hidden",(w=n.parentElement)===null||w===void 0||w.appendChild(F);var C=u.current===p?a:void 0,A=m?_Y(y,t.current,F,C):EY(y,t.current,F,C);(k=n.parentElement)===null||k===void 0||k.removeChild(F),!a&&A||a&&A&&!HY(a,A)&&l.current<5?(l.current++,s(A)):l.current>0&&(l.current=0,_==null||_(a))}},n);return u.current=p,function(){d.cancelAnimationFrame(E),u.current=void 0}}},[h,b,d,n,v,t,r,m,o,_,a,e,p]),a}function LY(e,t,n){var r=e.hidden,o=e.setInitialFocus,i=Zd(),a=!!t;T.useEffect(function(){if(!r&&o&&a&&n){var s=i.requestAnimationFrame(function(){return iK(n)},n);return function(){return i.cancelAnimationFrame(s)}}},[r,a,i,n,o])}function jY(e,t,n,r,o){var i=e.hidden,a=e.onDismiss,s=e.preventDismissOnScroll,l=e.preventDismissOnResize,u=e.preventDismissOnLostFocus,d=e.dismissOnTargetClick,h=e.shouldDismissOnWindowFocus,p=e.preventDismissOnEvent,m=T.useRef(!1),v=Zd(),_=A4([function(){m.current=!0},function(){m.current=!1}]),b=!!t;return T.useEffect(function(){var E=function(A){b&&!s&&y(A)},w=function(A){!l&&!(p&&p(A))&&(a==null||a(A))},k=function(A){u||y(A)},y=function(A){var P=A.composedPath?A.composedPath():[],I=P.length>0?P[0]:A.target,j=n.current&&!bE(n.current,I);if(j&&m.current){m.current=!1;return}if(!r.current&&j||A.target!==o&&j&&(!r.current||"stopPropagation"in r.current||d||I!==r.current&&!bE(r.current,I))){if(p&&p(A))return;a==null||a(A)}},F=function(A){h&&(p&&!p(A)||!p&&!u)&&!(o!=null&&o.document.hasFocus())&&A.relatedTarget===null&&(a==null||a(A))},C=new Promise(function(A){v.setTimeout(function(){if(!i&&o){var P=[zf(o,"scroll",E,!0),zf(o,"resize",w,!0),zf(o.document.documentElement,"focus",k,!0),zf(o.document.documentElement,"click",k,!0),zf(o,"blur",F,!0)];A(function(){P.forEach(function(I){return I()})})}},0)});return function(){C.then(function(A){return A()})}},[i,v,n,r,o,a,h,d,u,l,s,b,p]),_}var aI=T.memo(T.forwardRef(function(e,t){var n=S4(iI,e),r=n.styles,o=n.style,i=n.ariaLabel,a=n.ariaDescribedBy,s=n.ariaLabelledBy,l=n.className,u=n.isBeakVisible,d=n.children,h=n.beakWidth,p=n.calloutWidth,m=n.calloutMaxWidth,v=n.calloutMinWidth,_=n.doNotLayer,b=n.finalHeight,E=n.hideOverflow,w=E===void 0?!!b:E,k=n.backgroundColor,y=n.calloutMaxHeight,F=n.onScroll,C=n.shouldRestoreFocus,A=C===void 0?!0:C,P=n.target,I=n.hidden,j=n.onLayerMounted,H=T.useRef(null),K=T.useState(null),U=K[0],pe=K[1],se=T.useCallback(function(tr){pe(tr)},[]),J=G0(H,t),$=rI(n.target,{current:U}),_e=$[0],ve=$[1],fe=DY(n,_e,ve),R=MY(n,H,U,_e,fe),L=PY(n,fe,R),Ae=jY(n,R,H,_e,ve),Ue=Ae[0],Ve=Ae[1],Le=(R==null?void 0:R.elementPosition.top)&&(R==null?void 0:R.elementPosition.bottom),st=oe(oe({},R==null?void 0:R.elementPosition),{maxHeight:L});if(Le&&(st.bottom=void 0),LY(n,R,U),T.useEffect(function(){I||j==null||j()},[I]),!ve)return null;var We=w,rt=u&&!!P,Zt=OY(r,{theme:n.theme,className:l,overflowYHidden:We,calloutWidth:p,positions:R,beakWidth:h,backgroundColor:k,calloutMaxWidth:m,calloutMinWidth:v,doNotLayer:_}),qn=oe(oe({maxHeight:y||"100%"},o),We&&{overflowY:"hidden"}),er=n.hidden?{visibility:"hidden"}:void 0;return T.createElement("div",{ref:J,className:Zt.container,style:er},T.createElement("div",oe({},Zr(n,W0,RY),{className:Ys(Zt.root,R&&R.targetEdge&&IY[R.targetEdge]),style:R?oe({},st):BY,tabIndex:-1,ref:se}),rt&&T.createElement("div",{className:Zt.beak,style:zY(R)}),rt&&T.createElement("div",{className:Zt.beakCurtain}),T.createElement(oI,{role:n.role,"aria-roledescription":n["aria-roledescription"],ariaDescribedBy:a,ariaLabel:i,ariaLabelledBy:s,className:Zt.calloutMain,onDismiss:n.onDismiss,onMouseDown:Ue,onMouseUp:Ve,onRestoreFocus:n.onRestoreFocus,onScroll:F,shouldRestoreFocus:A,style:qn},d)))}),function(e,t){return!t.shouldUpdateWhenHidden&&e.hidden&&t.hidden?!0:T4(e,t)});function zY(e){var t,n,r=oe(oe({},(t=e==null?void 0:e.beakPosition)===null||t===void 0?void 0:t.elementPosition),{display:!((n=e==null?void 0:e.beakPosition)===null||n===void 0)&&n.hideBeak?"none":void 0});return!r.top&&!r.bottom&&!r.left&&!r.right&&(r.left=xx.left,r.top=xx.top),r}function HY(e,t){return Cx(e.elementPosition,t.elementPosition)&&Cx(e.beakPosition.elementPosition,t.beakPosition.elementPosition)}function Cx(e,t){for(var n in t)if(t.hasOwnProperty(n)){var r=e[n],o=t[n];if(r!==void 0&&o!==void 0){if(r.toFixed(2)!==o.toFixed(2))return!1}else return!1}return!0}aI.displayName=FY;function UY(e){return{height:e,width:e}}var qY={container:"ms-Callout-container",root:"ms-Callout",beak:"ms-Callout-beak",beakCurtain:"ms-Callout-beakCurtain",calloutMain:"ms-Callout-main"},$Y=function(e){var t,n=e.theme,r=e.className,o=e.overflowYHidden,i=e.calloutWidth,a=e.beakWidth,s=e.backgroundColor,l=e.calloutMaxWidth,u=e.calloutMinWidth,d=e.doNotLayer,h=hs(qY,n),p=n.semanticColors,m=n.effects;return{container:[h.container,{position:"relative"}],root:[h.root,n.fonts.medium,{position:"absolute",display:"flex",zIndex:d?Dd.Layer:void 0,boxSizing:"border-box",borderRadius:m.roundedCorner2,boxShadow:m.elevation16,selectors:(t={},t[Ao]={borderWidth:1,borderStyle:"solid",borderColor:"WindowText"},t)},GV(),r,!!i&&{width:i},!!l&&{maxWidth:l},!!u&&{minWidth:u}],beak:[h.beak,{position:"absolute",backgroundColor:p.menuBackground,boxShadow:"inherit",border:"inherit",boxSizing:"border-box",transform:"rotate(45deg)"},UY(a),s&&{backgroundColor:s}],beakCurtain:[h.beakCurtain,{position:"absolute",top:0,right:0,bottom:0,left:0,backgroundColor:p.menuBackground,borderRadius:m.roundedCorner2}],calloutMain:[h.calloutMain,{backgroundColor:p.menuBackground,overflowX:"hidden",overflowY:"auto",position:"relative",width:"100%",borderRadius:m.roundedCorner2},o&&{overflowY:"hidden"},s&&{backgroundColor:s}]}},WY=gl(aI,$Y,void 0,{scope:"CalloutContent"}),sI={exports:{}},ea={},lI={exports:{}},uI={};/** @license React v0.20.2
  * scheduler.production.min.js
  *
